### 2022-09-07 #4745 @editorialbot
[PRE REVIEW]: simcardems: A FEniCS-based cardiac electro-mechanics solver
**Submitting author:** @finsberg (<a href="http://orcid.org/0000-0003-3766-2393">Henrik Nicolay Topnes Finsberg</a>)
**Repository:** https://github.com/ComputationalPhysiology/simcardems
**Branch with paper.md** (empty if default branch): master
**Version:** v2022.3.0
