+++ joss/paper.bib
@@ -7,6 +7,7 @@
 }
 
 @article{finsberg2019pulse,
+  doi     = {10.21105/joss.01539},
   title   = {pulse: A python package based on FEniCS for solving problems in cardiac mechanics},
   author  = {Finsberg, Henrik Nicolay Topnes},
   journal = {Journal of Open Source Software},
@@ -17,6 +18,7 @@
 }
 
 @article{rognes2017cbcbeat,
+  doi     = {10.21105/joss.00224},
   title   = {cbcbeat: an adjoint-enabled framework for computational cardiac electrophysiology},
   author  = {Rognes, Marie E and Farrell, Patrick E and Funke, Simon W and Hake, Johan E and others},
   journal = {Journal of Open Source Software},
@@ -27,6 +29,7 @@
 }
 
 @article{arens2018gems,
+  doi       = {10.3389/fphys.2018.01431},
   title     = {GEMS: A fully integrated PETSc-based solver for coupled cardiac electromechanics and bidomain simulations},
   author    = {Arens, Sander and Dierckx, Hans and Panfilov, Alexander V},
   journal   = {Frontiers in physiology},
@@ -61,6 +64,7 @@
 }
 
 @article{land2017model,
+  doi={10.1016/j.yjmcc.2017.03.008},
   title={A model of cardiac contraction based on novel measurements of tension development in human cardiomyocytes},
   author={Land, Sander and Park-Holohan, So-Jin and Smith, Nicolas P and Dos Remedios, Cristobal G and Kentish, Jonathan C and Niederer, Steven A},
   journal={Journal of molecular and cellular cardiology},
