### 2022-09-11 #4745 @finsberg
> Ok, what's the submission about?

@cmorris2945  You can read more about the submission here: https://github.com/ComputationalPhysiology/simcardems but in short it is a library for performing cardiac electro-mechanics simulations
