### 2022-09-12 #4745 @cmorris2945
> > Ok, what's the submission about?
>
> @cmorris2945 You can read more about the submission here: https://github.com/ComputationalPhysiology/simcardems but in short it is a library for performing cardiac electro-mechanics simulations

> > Ok, what's the submission about?
>
> @cmorris2945 You can read more about the submission here: https://github.com/ComputationalPhysiology/simcardems but in short it is a library for performing cardiac electro-mechanics simulations

ok thanks
