+++ src/simcardems/geometry.py
@@ -10,6 +10,14 @@ from . import utils
 logger = utils.getLogger(__name__)
 
 
+def box_marker(p0, p1, x, y, z):
+    eps = dolfin.DOLFIN_EPS
+    in_xrange = p0[0] - eps < x < p1[0] + eps
+    in_yrange = p0[1] - eps < y < p1[1] + eps
+    in_zrange = p0[2] - eps < z < p1[2] + eps
+    return in_xrange and in_yrange and in_zrange
+
+
 def refine_mesh(
     mesh: dolfin.Mesh,
     num_refinements: int,
@@ -49,12 +57,25 @@ class BaseGeometry(abc.ABC):
 
     num_refinements: int = 0
     mechanics_mesh: dolfin.Mesh
+    mechanics_ht_mesh: dolfin.Mesh = None
 
+    # ep_mesh : Return heart mesh (refined) used in EP model
     @property
     def ep_mesh(self) -> dolfin.Mesh:
         if not hasattr(self, "_ep_mesh"):
             self._ep_mesh = refine_mesh(self.mechanics_mesh, self.num_refinements)
         return self._ep_mesh
+
+    # ep_mesh : Return full (heart + torso) mesh used in EP model
+    @property
+    def ep_ht_mesh(self) -> dolfin.Mesh:
+        if self.mechanics_ht_mesh is None:
+            return None
+        if not hasattr(self, "_ep_ht_mesh"):
+            self._ep_ht_mesh = self.mechanics_ht_mesh
+            for i in range(self.num_refinements):
+                self._ep_ht_mesh = dolfin.adapt(self._ep_ht_mesh)
+        return self._ep_ht_mesh
 
     @abc.abstractproperty
     def parameters(self) -> Dict[str, float]:
@@ -107,6 +128,109 @@ class SlabGeometry(BaseGeometry):
             # For example we should make sure that num_refinements are correct
             # and the the ep mesh has the same partition as the mechanics mesh
             self._ep_mesh = ep_mesh
+
+    def __repr__(self) -> str:
+        return (
+            f"{self.__class__.__name__}("
+            f"lx={self.lx}, "
+            f"ly={self.ly}, "
+            f"lz={self.lz}, "
+            f"dx={self.dx}, "
+            f"num_refinements={self.num_refinements})"
+        )
+
+    @property
+    def parameters(self):
+        return {
+            "lx": self.lx,
+            "ly": self.ly,
+            "lz": self.lz,
+            "dx": self.dx,
+            "num_refinements": self.num_refinements,
+        }
+
+
+class TorsoGeometry(BaseGeometry):
+    def __init__(
+        self,
+        lx: float,
+        ly: float,
+        lz: float,
+        dx: float,
+        num_refinements: int = 0,
+        mechanics_mesh: Optional[dolfin.Mesh] = None,
+        mechanics_heart_marker: Optional[dolfin.MeshFunction] = None,
+    ) -> None:
+        self.lx = lx
+        self.ly = ly
+        self.lz = lz
+        self.dx = dx
+        self.num_refinements = num_refinements
+
+        # Build torso mesh
+        if mechanics_mesh is None:
+            N = lambda v: int(np.rint(v))
+            mechanics_mesh = dolfin.BoxMesh(
+                dolfin.MPI.comm_world,
+                dolfin.Point(-0.5 * lx, -0.5 * ly, -0.5 * lz),
+                dolfin.Point(1.5 * lx, 1.5 * ly, 1.5 * lz),
+                N(2 * lx / dx),
+                N(2 * ly / dx),
+                N(2 * lz / dx),
+            )
+        else:
[…]
