### 2022-09-19 087dbec
Setup EP model for the torso (TorsoModel) when setting up EP model - Important note : Requires the appropriate branch in cbcbeat (coupling-heart-torso)
