### 2022-09-19 8048821
Add mechanics torso model (simplest possible - Poisson with appropriate displacement BC at inner boundary) - Needs some fixes (way too expensive for now)
