+++ src/simcardems/mechanics_model.py
@@ -211,6 +211,87 @@ class RigidMotionProblem(MechanicsProblem):
         return sum(dolfin.dot(u, zi) * r[i] * dolfin.dx for i, zi in enumerate(RM))
 
 
+class TorsoMechanicsProblem:
+    def __init__(self, full_mesh, torso_mesh, heart_d):
+        self.mesh = full_mesh
+        self.torso_mesh = torso_mesh
+        self.heart_mesh = heart_d.function_space().mesh()
+        self.heart_d = heart_d
+        self.V = dolfin.FunctionSpace(self.torso_mesh, self.heart_d.ufl_element())
+
+    def update_heart_disp(self, d):
+        self.heart_d = d
+
+    def set_bcs(self):
+        bc_markers = torso_boundary_marking(self.torso_mesh, self.heart_mesh)
+        with dolfin.XDMFFile(
+            dolfin.MPI.comm_world,
+            "./XDMF/HT/torso_marking.xdmf",
+        ) as xdmf:
+            xdmf.write(bc_markers)
+
+        ## FIXME : This is super expensive!
+        class HeartDisp(dolfin.UserExpression):
+            def __init__(self, mesh, heart_d, **kwargs):
+                self.mesh_ = mesh
+                self.heart_disp_ = heart_d
+                super().__init__(**kwargs)
+
+            def eval(self, values, x):
+                # Check if x is in heart mesh (should be True only for heart boundary points)
+                index = self.mesh_.bounding_box_tree().compute_first_entity_collision(
+                    dolfin.Point(x),
+                )
+                x_in_mesh = index < self.mesh_.num_entities(self.mesh_.topology().dim())
+                if x_in_mesh:
+                    for i in range(self.mesh_.topology().dim()):
+                        values[i] = self.heart_disp_(x)[i]
+
+            def value_shape(self):
+                return (self.mesh_.topology().dim(),)
+
+        zero = dolfin.Constant((0.0, 0.0, 0.0))
+        dh = HeartDisp(self.heart_mesh, self.heart_d, degree=2)
+
+        # DEBUG
+        # Interpolation is also super expensive!
+        # dh_h = dolfin.interpolate(dh, self.V)
+        # with dolfin.XDMFFile(dolfin.MPI.comm_world, "./XDMF/HT/heart_disp1.xdmf") as xdmf:
+        #     xdmf.write(dh_h)
+        # with dolfin.XDMFFile(dolfin.MPI.comm_world, "./XDMF/HT/heart_disp2.xdmf") as xdmf:
+        #     xdmf.write(self.heart_d)
+
+        bc_ext = dolfin.DirichletBC(self.V, zero, bc_markers, 2)
+        bc_int = dolfin.DirichletBC(self.V, dh, bc_markers, 1)
+
+        return [bc_int, bc_ext]
+
+    def solve(self):
+
+        u = dolfin.TrialFunction(self.V)
+        phi = dolfin.TestFunction(self.V)
+        d = dolfin.Function(self.V)
+
+        # Variational formulation - Standard Poisson (for now)
+        a = dolfin.inner(dolfin.grad(u), dolfin.grad(phi)) * dolfin.dx
+        L = dolfin.inner(dolfin.Constant((0.0, 0.0, 0.0)), phi) * dolfin.dx
+
+        bcs = self.set_bcs()
[…]
