+++ src/simcardems/setup_models.py
@@ -22,21 +22,30 @@ logger = utils.getLogger(__name__)
 
 EMState = namedtuple(
     "EMState",
-    ["coupling", "solver", "mech_heart", "t0"],
+    ["coupling", "solver", "mech_heart", "mech_torso", "t0"],
 )
 
 
 def setup_EM_model(config: Config):
 
-    geo = geometry.SlabGeometry(
-        lx=config.lx,
-        ly=config.ly,
-        lz=config.lz,
-        dx=config.dx,
-        num_refinements=config.num_refinements,
-    )
-
-    coupling = em_model.EMCoupling(geo)
+    if config.heart_torso_coupling:
+        geo_t = geometry.TorsoGeometry(
+            lx=config.lx,
+            ly=config.ly,
+            lz=config.lz,
+            dx=config.dx,
+            num_refinements=config.num_refinements,
+        )
+        coupling = em_model.EMCoupling(geo_t)
+    else:
+        geo_s = geometry.SlabGeometry(
+            lx=config.lx,
+            ly=config.ly,
+            lz=config.lz,
+            dx=config.dx,
+            num_refinements=config.num_refinements,
+        )
+        coupling = em_model.EMCoupling(geo_s)
 
     # Set-up solver and time it
     solver = setup_ep_solver(
@@ -51,7 +60,7 @@ def setup_EM_model(config: Config):
 
     coupling.register_ep_model(solver)
 
-    mech_heart = setup_mechanics_solver(
+    (mech_heart, mech_torso) = setup_mechanics_solver(
         coupling=coupling,
         bnd_cond=config.bnd_cond,
         cell_params=solver.ode_solver._model.parameters(),
@@ -62,12 +71,14 @@ def setup_EM_model(config: Config):
         set_material=config.set_material,
         mechanics_ode_scheme=config.mechanics_ode_scheme,
         use_custom_newton_solver=config.mechanics_use_custom_newton_solver,
+        heart_torso_coupling=config.heart_torso_coupling,
     )
 
     return EMState(
         coupling=coupling,
         solver=solver,
         mech_heart=mech_heart,
+        mech_torso=mech_torso,
         t0=0,
     )
 
@@ -84,6 +95,7 @@ def setup_mechanics_solver(
     set_material: str = "",
     linear_solver="mumps",
     use_custom_newton_solver: bool = Config.mechanics_use_custom_newton_solver,
+    heart_torso_coupling: bool = Config.heart_torso_coupling,
     Zetas_prev=None,
     Zetaw_prev=None,
     lmbda_prev=None,
@@ -179,7 +191,17 @@ def setup_mechanics_solver(
     logger.info("Mechanics model")
     utils.print_mesh_info(coupling.mech_mesh, total_dofs)
 
-    return problem
+    torso_problem = None
+    if heart_torso_coupling:
+        u, p = problem.state.split(deepcopy=True)
+        # u, p = problem.state.split()
+        # print("displacement min = ", min(u.vector().get_local()))
+        # print("displacement max = ", max(u.vector().get_local()))
+        TorsoProblem = mechanics_model.TorsoMechanicsProblem
+        torso_problem = TorsoProblem(coupling.mech_ht_mesh, coupling.mech_torso_mesh, u)
+        torso_problem.solve()
+
+    return (problem, torso_problem)
 
 
 def setup_ep_solver(
@@ -195,6 +217,7 @@ def setup_ep_solver(
     popu_factors_file=None,
     disease_state=Config.disease_state,
     PCL=Config.PCL,
+    heart_torso_coupling=Config.heart_torso_coupling,
 ):
     ps = ep_model.setup_splitting_solver_parameters(
         theta=theta,
@@ -222,10 +245,16 @@ def setup_ep_solver(
[…]
