### 2022-09-26 102c38a
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/asottile/reorder_python_imports: v3.8.2 → v3.8.3](https://github.com/asottile/reorder_python_imports/compare/v3.8.2...v3.8.3)
- [github.com/asottile/add-trailing-comma: v2.2.3 → v2.3.0](https://github.com/asottile/add-trailing-comma/compare/v2.2.3...v2.3.0)
- [github.com/streetsidesoftware/cspell-cli: v6.8.1 → v6.10.0](https://github.com/streetsidesoftware/cspell-cli/compare/v6.8.1...v6.10.0)
