+++ src/simcardems/land_model.py
@@ -43,7 +43,10 @@ class LandModel(pulse.ActiveModel):
     ):
         super().__init__(f0=f0, s0=s0, n0=n0)
         self._eta = eta
-        self.function_space = dolfin.FunctionSpace(mesh, "CG", 1)
+        # Use DG 1 rep of lambda and zetas?
+        self.quad_space = pulse.QuadratureSpace(mesh, degree=3, dim=1)
+        self.cg1_space = dolfin.FunctionSpace(mesh, "CG", 1)
+        # self.dg1_space = dolfin.FunctionSpace(mesh, "DG", 1)
 
         self.XS = XS
         self.XW = XW
@@ -52,24 +55,24 @@ class LandModel(pulse.ActiveModel):
         self._t_prev = 0.0
         self._scheme = scheme
 
-        self._dLambda = dolfin.Function(self.function_space)
-        self.lmbda_prev = dolfin.Function(self.function_space)
+        self._dLambda = dolfin.Function(self.quad_space)
+        self.lmbda_prev = dolfin.Function(self.quad_space)
         self.lmbda_prev.vector()[:] = 1.0
         if lmbda is not None:
             self.lmbda_prev.assign(lmbda)
-        self.lmbda = dolfin.Function(self.function_space)
+        self.lmbda = dolfin.Function(self.quad_space)
 
-        self._Zetas = dolfin.Function(self.function_space)
-        self.Zetas_prev = dolfin.Function(self.function_space)
+        self._Zetas = dolfin.Function(self.quad_space)
+        self.Zetas_prev = dolfin.Function(self.quad_space)
         if Zetas is not None:
             self.Zetas_prev.assign(Zetas)
 
-        self._Zetaw = dolfin.Function(self.function_space)
-        self.Zetaw_prev = dolfin.Function(self.function_space)
+        self._Zetaw = dolfin.Function(self.quad_space)
+        self.Zetaw_prev = dolfin.Function(self.quad_space)
         if Zetaw is not None:
             self.Zetaw_prev.assign(Zetaw)
 
-        self.Ta_current = dolfin.Function(self.function_space, name="Ta")
+        self.Ta_current = dolfin.Function(self.cg1_space, name="Ta")
 
     @property
     def dLambda(self):
@@ -178,8 +181,9 @@ class LandModel(pulse.ActiveModel):
     def update_prev(self):
         self.Zetas_prev.vector()[:] = self.Zetas.vector()
         self.Zetaw_prev.vector()[:] = self.Zetaw.vector()
+        # Use previous displacements
         self.lmbda_prev.vector()[:] = self.lmbda.vector()
-        self.Ta_current.assign(dolfin.project(self.Ta, self.function_space))
+        self.Ta_current.assign(dolfin.project(self.Ta, self.cg1_space))
 
     @property
     def Ta(self):
@@ -209,7 +213,16 @@ class LandModel(pulse.ActiveModel):
 
         C = F.T * F
         f = F * self.f0
-        self.lmbda.assign(dolfin.project(dolfin.sqrt(f**2), self.function_space))
+
+        # self.lmbda.assign(dolfin.project(dolfin.sqrt(f**2), self.function_space))
+
+        self.lmbda.assign(
+            dolfin.project(
+                dolfin.sqrt(f**2),
+                self.quad_space,
+                form_compiler_parameters={"representation": "quadrature"},
+            ),
+        )
         self.update_Zetas()
         self.update_Zetaw()
 
