### 2022-10-03 94e796a
Project lambda and zetas to CG 1 space on the fine mesh so that we can use the PointIntegralSolver in CBCbeat
