+++ src/simcardems/em_model.py
@@ -1,6 +1,7 @@
 import dolfin
 
 from . import geometry as _geometry
+from . import land_model
 from . import utils
 from .mechanics_model import MechanicsProblem
 
@@ -19,15 +20,24 @@ class EMCoupling:
         self.XS_mech = dolfin.Function(self.V_mech, name="XS_mech")
         self.XW_mech = dolfin.Function(self.V_mech, name="XW_mech")
         self.lmbda_mech = dolfin.Function(self.V_mech, name="lambda_mech")
-        self.Zetas_mech = dolfin.Function(self.V_mech, name="Zetas_mech")
-        self.Zetaw_mech = dolfin.Function(self.V_mech, name="Zetaw_mech")
+        # self.Zetas_mech = dolfin.Function(self.V_mech, name="Zetas_mech")
+        # self.Zetaw_mech = dolfin.Function(self.V_mech, name="Zetaw_mech")
 
         self.V_ep = dolfin.FunctionSpace(self.ep_mesh, "CG", 1)
         self.XS_ep = dolfin.Function(self.V_ep, name="XS_ep")
         self.XW_ep = dolfin.Function(self.V_ep, name="XW_ep")
-        self.lmbda_ep = dolfin.Function(self.V_ep, name="lambda_ep")
-        self.Zetas_ep = dolfin.Function(self.V_ep, name="Zetas_ep")
-        self.Zetaw_ep = dolfin.Function(self.V_ep, name="Zetaw_ep")
+
+        # self.Q_ep = pulse.QuadratureSpace(self.ep_mesh, degree=3, dim=1)
+        self.Q_ep = self.V_ep  # This should eventually be a quadrature space
+        self.lmbda_ep = dolfin.Function(self.Q_ep, name="lambda_ep")
+        self.lmbda_ep_prev = dolfin.Function(self.Q_ep, name="lambda_ep_prev")
+        self.lmbda_ep_prev.vector()[:] = 1.0
+        self._dLambda_ep = dolfin.Function(self.Q_ep, name="lambda_ep")
+        self.Zetas_ep = dolfin.Function(self.Q_ep, name="Zetas_ep")
+        self.Zetaw_ep = dolfin.Function(self.Q_ep, name="Zetaw_ep")
+        self.Zetas_ep_prev = dolfin.Function(self.Q_ep, name="Zetas_ep_prev")
+        self.Zetaw_ep_prev = dolfin.Function(self.Q_ep, name="Zetaw_ep_prev")
+
         self.W_ep = dolfin.VectorFunctionSpace(self.ep_mesh, "CG", 2)
         self.u_ep = dolfin.Function(self.W_ep)
 
@@ -38,6 +48,13 @@ class EMCoupling:
     @property
     def ep_mesh(self):
         return self.geometry.ep_mesh
+
+    @property
+    def dLambda(self):
+        self._dLambda_ep.vector()[:] = (
+            self.lmbda_ep.vector() - self.lmbda_ep_prev.vector()
+        )
+        return self._dLambda_ep
 
     def register_ep_model(self, solver):
         logger.debug("Registering EP model")
@@ -59,12 +76,16 @@ class EMCoupling:
         )
         self.f0 = solver.material.f0
 
-        self.Zetas_mech = solver.material.active.Zetas_prev
-        self.Zetaw_mech = solver.material.active.Zetaw_prev
+        # self.Zetas_mech = solver.material.active.Zetas_prev
+        # self.Zetaw_mech = solver.material.active.Zetaw_prev
         # self.lmbda_mech = solver.material.active.lmbda_prev
-        self.mech_state = solver.state
+        self.mech_solver = solver
         self.mechanics_to_coupling()
         logger.debug("Done registering EP model")
[…]
