### 2022-10-03 2e030bd
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pre-commit/mirrors-mypy: v0.971 → v0.981](https://github.com/pre-commit/mirrors-mypy/compare/v0.971...v0.981)
- [github.com/streetsidesoftware/cspell-cli: v6.10.0 → v6.12.0](https://github.com/streetsidesoftware/cspell-cli/compare/v6.10.0...v6.12.0)
