+++ src/simcardems/em_model.py
@@ -19,9 +19,6 @@ class EMCoupling:
         self.V_mech = dolfin.FunctionSpace(self.mech_mesh, "CG", 1)
         self.XS_mech = dolfin.Function(self.V_mech, name="XS_mech")
         self.XW_mech = dolfin.Function(self.V_mech, name="XW_mech")
-        self.lmbda_mech = dolfin.Function(self.V_mech, name="lambda_mech")
-        # self.Zetas_mech = dolfin.Function(self.V_mech, name="Zetas_mech")
-        # self.Zetaw_mech = dolfin.Function(self.V_mech, name="Zetaw_mech")
 
         self.V_ep = dolfin.FunctionSpace(self.ep_mesh, "CG", 1)
         self.XS_ep = dolfin.Function(self.V_ep, name="XS_ep")
@@ -75,10 +72,6 @@ class EMCoupling:
             self._u_subspace_index,
         )
         self.f0 = solver.material.f0
-
-        # self.Zetas_mech = solver.material.active.Zetas_prev
-        # self.Zetaw_mech = solver.material.active.Zetaw_prev
-        # self.lmbda_mech = solver.material.active.lmbda_prev
         self.mech_solver = solver
         self.mechanics_to_coupling()
         logger.debug("Done registering EP model")
