### 2022-10-06 b909fbf
Try to add workflow to run benchmark
