+++ src/simcardems/postprocess.py
@@ -1,15 +1,17 @@
 import json
 import warnings
+from enum import Enum
 from pathlib import Path
 
-import ap_features as apf
 import dolfin
 import h5py
 import matplotlib.pyplot as plt
 import numpy as np
 import tqdm
 
+import ap_features as apf
 from . import utils
+from .datacollector import DataGroups
 from .datacollector import DataLoader
 
 logger = utils.getLogger(__name__)
@@ -19,29 +21,31 @@ def center_func(fmin, fmax):
     return fmin + (fmax - fmin) / 2
 
 
+class BoundaryNodes(Enum):
+    center = "center"
+    node_A = "node_A"
+    node_B = "node_B"
+    node_C = "node_C"
+    node_D = "node_D"
+    node_E = "node_E"
+    node_F = "node_F"
+    node_G = "node_G"
+    node_H = "node_H"
+    xmax = "xmax"
+    xmin = "xmin"
+    ymax = "ymax"
+    ymin = "ymin"
+    zmax = "zmax"
+    zmin = "zmin"
+
+
 class Boundary:
     def __init__(self, mesh):
         self.mesh = mesh
 
     @staticmethod
     def nodes():
-        return (
-            "center",
-            "node_A",
-            "node_B",
-            "node_C",
-            "node_D",
-            "node_E",
-            "node_F",
-            "node_G",
-            "node_H",
-            "xmax",
-            "xmin",
-            "ymax",
-            "ymin",
-            "zmax",
-            "zmin",
-        )
+        return BoundaryNodes._member_names_
 
     @property
     def boundaries(self):
@@ -281,6 +285,62 @@ def plot_peaks(fname, data, threshold):
     fig.savefig(fname, dpi=300)
 
 
+def extract_traces(loader: DataLoader, point: BoundaryNodes = BoundaryNodes.center):
+    bnd = {"ep": Boundary(loader.ep_mesh), "mechanics": Boundary(loader.mech_mesh)}
+
+    all_names = {
+        "mechanics": [
+            "lmbda",
+            "Ta",
+            "Zetas_mech",
+            "Zetaw_mech",
+            "XS_mech",
+            "XW_mech",
+            "u",
+        ],
+        "ep": ["V", "Ca", "XS", "XW", "CaTrpn", "TmB", "Cd", "Zetas", "Zetaw"],
+    }
+
+    values = {
+        group: {
+            name: np.zeros(len(loader.time_stamps)) for name in names if name != "u"
+        }
+        for group, names in all_names.items()
+    }
+
+    values["mechanics"]["u"] = np.zeros((len(loader.time_stamps), 3))
+    values["time"] = np.array(loader.time_stamps, dtype=float)
+
+    for i, t in enumerate(loader.time_stamps):
+
+        for group, names in all_names.items():
+            value_point = getattr(bnd[group], utils.enum2str(point, BoundaryNodes))
+            datagroup = getattr(DataGroups, utils.enum2str(group, DataGroups))
+            for name in names:
+                func = loader.get(datagroup, name, t)
+
+                dof_coords = func.function_space().tabulate_dof_coordinates()
+                dof = np.argmin(
+                    np.linalg.norm(dof_coords - np.array(value_point), axis=1),
+                )
+                if name == "u":
+                    # e.g u -> a vector
+                    values[group][name][i] = func(value_point)
+
[…]
