+++ .github/workflows/benchmark.yml
@@ -31,6 +31,6 @@ jobs:
         run: |
           git config --global user.name "${{ env.CI_COMMIT_AUTHOR }}"
           git config --global user.email "username@users.noreply.github.com"
-          find benchmarks -name '*.json' | xargs git add
+          find benchmarks -name '*.json' | xargs git add -f
           git commit -a -m "${{ env.CI_COMMIT_MESSAGE }}"
           git push
