### 2022-10-06 179963b
More enum fixes
