### 2022-10-06 a73298f
Run full benchmark
