### 2022-10-06 7d375ab
Only run benchmark on new tags
