### 2022-10-06 3fbc28b
Add timestamp and version to results and use git hash for folder
