+++ .github/workflows/benchmark.yml
@@ -8,9 +8,7 @@ name: Run benchmark
 on:
   push:
     branches:
-      - "!*"
-    tags:
-      - "v*"
+      - master
 env:
   CI_COMMIT_MESSAGE: Run benchmark and upload report
   CI_COMMIT_AUTHOR: Continuous Integration
@@ -29,7 +27,7 @@ jobs:
         run: python -m pip install -e "."
 
       - name: Run benchmark
-        run: python -m simcardems run-benchmark benchmarks
+        run: python -m simcardems run-benchmark "benchmarks/${{ github.sha }}"
 
       - name: Push benchmark report
         run: |
@@ -37,4 +35,6 @@ jobs:
           git config --global user.email "username@users.noreply.github.com"
           find benchmarks -name '*.json' | xargs git add -f
           git commit -a -m "${{ env.CI_COMMIT_MESSAGE }}"
+          git config pull.rebase true
+          git pull
           git push
