### 2022-10-06 1b5557a
Add benchmark run command
