+++ .github/workflows/benchmark.yml
@@ -27,7 +27,7 @@ jobs:
         run: python -m pip install -e "."
 
       - name: Run benchmark
-        run: python -m simcardems run-benchmark "benchmarks/${{ github.sha }}"
+        run: make run-benchmark
 
       - name: Push benchmark report
         run: |
