+++ .github/workflows/benchmark.yml
@@ -24,7 +24,10 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Install dependencies
-        run: python -m pip install -e "."
+        run: |
+          python -m pip install -e "."
+          python -m pip install pre-commit
+          pre-commit install
 
       - name: Run benchmark
         run: make run-benchmark
