+++ .github/workflows/benchmark.yml
@@ -26,8 +26,6 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install -e "."
-          python -m pip install pre-commit
-          pre-commit install
 
       - name: Run benchmark
         run: make run-benchmark
