### 2022-10-06 630ed09
Do not run pre-commit hooks on .json files
