+++ src/simcardems/postprocess.py
@@ -3,13 +3,13 @@ import warnings
 from enum import Enum
 from pathlib import Path
 
-import ap_features as apf
 import dolfin
 import h5py
 import matplotlib.pyplot as plt
 import numpy as np
 import tqdm
 
+import ap_features as apf
 from . import utils
 from .datacollector import DataGroups
 from .datacollector import DataLoader
