+++ .pre-commit-config.yaml
@@ -4,6 +4,7 @@ repos:
     hooks:
       - id: check-yaml
       - id: end-of-file-fixer
+        exclude: ^benchmarks
       - id: trailing-whitespace
       - id: check-docstring-first
       - id: debug-statements
