+++ setup.cfg
@@ -53,9 +53,8 @@ dev =
     twine
     wheel
 docs =
-    Sphinx
-    myst-parser
-    sphinx-book-theme
+    jupyter-book
+    jupytext
     sphinxcontrib-bibtex
 gui =
     fenics-plotly
