+++ docs/_config.yml
@@ -1,0 +1,40 @@
+# Book settings
+# Learn more at https://jupyterbook.org/customize/config.html
+
+title: simcardems
+author: Simula Research Laboratory
+logo: "logo.png"
+copyright: "2022"
+only_build_toc_files: true
+
+# Force re-execution of notebooks on each build.
+# See https://jupyterbook.org/content/execute.html
+execute:
+  execute_notebooks: cache
+  timeout: 3000
+
+# Information about where the book exists on the web
+repository:
+  url: https://github.com/ComputationalPhysiology/simcardems # Online location of your book
+  path_to_book: _sources # Optional path to your book, relative to the repository root
+  branch: gh-pages # Which branch of the repository should be used when creating links (optional)
+
+html:
+  use_issues_button: true
+  use_repository_button: true
+
+parse:
+  myst_enable_extensions:
+    - amsmath
+    - dollarmath
+    - linkify
+
+sphinx:
+  config:
+    bibtex_bibfiles: ["refs.bib"]
+
+  extra_extensions:
+    - "sphinx.ext.autodoc"
+    - "sphinx.ext.napoleon"
+    - "sphinx.ext.viewcode"
+    - "sphinxcontrib.bibtex"
