+++ _sources/theory.rst.txt
@@ -1,15 +1,0 @@
-Mathematical Theory
-===================
-
-In this section we will cover the necessary mathematical for electro-mechanics.
-
-Content
--------
-
-.. toctree::
-   :maxdepth: 1
-
-   mechanics_model
-   electrophysiology_model
-   cell_model
-   electro_mechanical_coupling
