+++ _sources/electrophysiology_model.md.txt
@@ -1,4 +1,0 @@
-(section:ep)=
-# Tissue level electrophysiology model
-
-We use the Monodomain model to model the electrophysiology at the tissue level. See <https://cbcbeat.readthedocs.io/en/latest/cbcbeat.html#module-cbcbeat.monodomainsolver> for more information
