+++ _sources/modules.rst.txt
@@ -1,7 +1,0 @@
-simcardems
-==========
-
-.. toctree::
-   :maxdepth: 4
-
-   simcardems

[…]
