+++ _sources/cell_model.md
@@ -1,0 +1,1423 @@
+(section:cell)=
+# Cellular model
+
+
+## ORdmm_Land_em_coupling
+This document is automatically generated using [gotran](https://pypi.org/project/gotran/)
+
+Model is based on the work from {cite}`o2011simulation`, {cite}`land2017model`
+
+### Parameters
+| Name                  | Value                    |
+| --------------------- | ------------------------ |
+| $scale_{\mathcal{I}}$ | $1.02$                   |
+| $scale_{IK1}$         | $1.41$                   |
+| $scale_{IKr}$         | $1.12$                   |
+| $scale_{IKs}$         | $1.65$                   |
+| $scale_{INaL}$        | $2.27$                   |
+| $celltype$            | $0$                      |
+| $cao$                 | $1.8$                    |
+| $ko$                  | $5.4$                    |
+| $nao$                 | $1.4\!\times\!10 ^{2}$   |
+| $F$                   | $9.65\!\times\!10 ^{4}$  |
+| $R$                   | $8.31\!\times\!10 ^{3}$  |
+| $T$                   | $3.1\!\times\!10 ^{2}$   |
+| $L$                   | $1\!\times\!10 ^{-2}$    |
+| $rad$                 | $110\!\times\!10 ^{-5}$  |
+| $Ahf$                 | $0.99$                   |
+| $GNa$                 | $3.1\!\times\!10 ^{1}$   |
+| $thL$                 | $2\!\times\!10 ^{2}$     |
+| $Gto$                 | $200\!\times\!10 ^{-4}$  |
+| $\delta_{epi}$        | $1$                      |
+| $Aff$                 | $0.6$                    |
+| $Kmn$                 | $200\!\times\!10 ^{-5}$  |
+| $k2n$                 | $1\!\times\!10 ^{3}$     |
+| $tjca$                | $7.5\!\times\!10 ^{1}$   |
+| $zca$                 | $2$                      |
+| $bt$                  | $4.75$                   |
+| $B_{0}$               | $2.3$                    |
+| $B_{1}$               | $-2.4$                   |
+| $Tot_{A}$             | $2.5\!\times\!10 ^{1}$   |
+| $Tref$                | $1.2\!\times\!10 ^{2}$   |
+| $Trpn_{50}$           | $0.35$                   |
+| $calib$               | $1$                      |
+| $cat_{50 ref}$        | $0.81$                   |
+| $dLambda$             | $0$                      |
+| $emcoupling$          | $1$                      |
+| $etal$                | $2\!\times\!10 ^{2}$     |
+| $etas$                | $2\!\times\!10 ^{1}$     |
+| $gammas$              | $850\!\times\!10 ^{-5}$  |
+| $gammaw$              | $0.61$                   |
+| $isacs$               | $0$                      |
+| $ktrpn$               | $1\!\times\!10 ^{-1}$    |
[…]
