+++ _sources/index.md
@@ -1,0 +1,24 @@
+# Cardiac MPS motion
+
+`simcardems` is a FEniCS-based cardiac electro-mechanics solver and is
+developed as a part of the [SimCardio Test project](https://www.simcardiotest.eu/wordpress). The solver depends
+on [pulse](https://github.com/ComputationalPhysiology/pulse) and [cbcbeat](https://github.com/ComputationalPhysiology/cbcbeat)
+
+## Source code
+The source code can be found on [GitHub](https://github.com/ComputationalPhysiology/simcardems)
+
+
+## Contents
+```{tableofcontents}
+```
+
+## License
+`simcardems` is licensed under the GNU LGPL, version 2.1 or (at your option) any later version.
+`simcardems` is Copyright (2011-2022) by the authors and Simula Research Laboratory.
+
+## Authors
+`simcardems` is developed at [Simula Research Laboratory](https://www.simula.no). The core developers are
+
+- Henrik Finsberg (henriknf@simula.no)
+- Ilsbeth van Herck (ilse@simula.no)
+- Cécile Daversin-Catty (cecile@simula.no)
