+++ _sources/electrophysiology_model.md
@@ -1,0 +1,4 @@
+(section:ep)=
+# Tissue level electrophysiology model
+
+We use the Monodomain model to model the electrophysiology at the tissue level. See <https://cbcbeat.readthedocs.io/en/latest/cbcbeat.html#module-cbcbeat.monodomainsolver> for more information

[…]
