+++ _sources/index.md
@@ -1,4 +1,4 @@
-# Cardiac MPS motion
+# Simula Cardiac Electro-Mechanics Solver
 
 `simcardems` is a FEniCS-based cardiac electro-mechanics solver and is
 developed as a part of the [SimCardio Test project](https://www.simcardiotest.eu/wordpress). The solver depends
