### 2022-10-07 #4753 @sdelandtsheer
Hello, I am having issues with the testing. With conda it seems that everything is set up except dolfin. I tried with Docker but my machine is refusing to cooperate although the image is pulled. I use windows 10, that might be the problem.
The project is interesting and the code seems to be of very good quality, however I am unable to run the tests and reproduce the plots.
I will be traveling over the next week and I don't have any more time to review this project, the install might be too complex for my IT skills. It might be a good idea to designate another reviewer if possible.
