### 2022-10-07 ComputationalPhysiology/simcardems#84 @mbarzegary
[JOSS Review] Contribution and authorship
Dear authors,

I think the list of authors needs clarification. Only 3 authors are listed on the documentation page while the paper has 5 authors. This is in line with the output of `git shortlog --summary` on the repo:

```
23  Cécile
8  Cécile Daversin-Catty
259  Henrik Finsberg
17  Ilse van Herck
66  IlsevanHerck
29  pre-commit-ci[bot]
```

Can you please explain the contribution of the other authors?

P.S. This issue is related to #joss-review/7453 https://github.com/openjournals/joss-reviews/issues/4753
