### 2022-10-07 ComputationalPhysiology/simcardems#85 @mbarzegary
[JOSS Review] Reproducibility and functionality
Dear authors,

Generally speaking, the software runs as described in the documentation, and the demos can be set up by following the instruction. However, there are minor fixes that can improve the quality of the submission:
* The about page is empty, which is the first thing the user sees when the GUI is loaded (it says “The SIMula CARDiac ElectroMechanics solver is ...”)
* After running the simulations, I couldn’t visualize the results in ParaView. The XDMF files are there but the XDMFReader cannot open them. I tried to open them using PV 5.8 and 5.9 in Windows and PV 5.7 in Linux, always got a “cannot load data” error. You may need to elaborate on the postprocessing steps in the documentation.
* It’s unclear what “node_A”, “node_B”, etc., means in the postprocessing section of the GUI. Which nodes are they referring to? This may be addressed in the documentation as well, as mentioned above.
* While checking the demo for the command-line interface, I noticed that `python3 -m simcardems run --end-time=100` works but `python3 -m simcardems run -T=100` doesn’t. It says “Error: Invalid value for '-T' / '--end-time': '=100' is not a valid float.”

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4753
