### 2022-10-07 ComputationalPhysiology/simcardems#86 @mbarzegary
[JOSS Review] Tests
Dear authors,

I see that the software has automated tests implemented using CI procedures in GitHub actions, but the provided manual test for checking the coupling doesn’t work properly. It runs without any error but always complains about `FileNotFoundError: File release_test_results/results.h5 does not exist`, no matter if the directory `release_test_results` exists or not. Also, it is not clear how the provided test checks if correct results are produced or not.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4753
