### 2022-10-07 ComputationalPhysiology/simcardems#87 @mbarzegary
[JOSS Review] Performance and HPC
Dear authors,

Since HPC is pointed out in the paper (by mentioning it as an advantage of the current implementation and leveraging FEniCS), I think it should be elaborated a bit in the documentation.

I couldn’t find anything regarding the parallel execution in the documentation. Are you relying on the user knowledge of FEniCS for this? If so, then the submission is useful for the FEniCS community only, limiting its potential to be used by broader users (I mentioned this as a problem for the installation documentation as well). I suggest that you explain how to run the code in parallel in the documentation. Also, what does the `hpc` argument do? It seems that it just turns off the progress bar, yes?

Furthermore, in my tests using 6 CPU cores (done using the demo example), the code runs slower when I call it with MPI. I guess that it may be related to the problem size, but anyway, it can be nice if you add another demo (maybe with a bigger problem size) for demonstrating the parallel execution and its benefits.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4753
