### 2022-10-07 ComputationalPhysiology/simcardems#88 @mbarzegary
[JOSS Review] Statement of need and the paper
Dear authors,

The paper is written well, but I believe it is too short and lacks essential parts to illustrate the necessity of the developed software. I suggest that you improve the statement of need part by adding a schematic showing where the software stands by considering the performed coupling.

Moreover, I think it’s necessary to clarify why such software is needed, for people who don’t know about the SimCardioTest project. The current explanation is very brief. How do you compare `simcardems` to other open-source tools such as SimVascular? What is the necessity of having another tool for a similar domain? Elaborating on these things can enrich the quality of the submission.

I think it is beneficial if you mention the limitations of the software (can be in the documentation instead). It seems that the tool works with rectangular domains only. No? if not, how should the user set up a simulation with a more complex domain? Is the software for educational purposes only?

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4753
