### 2022-10-07 ComputationalPhysiology/simcardems#89 @mbarzegary
[JOSS Review] Documentation 
Dear authors,

The documentation is quite acceptable for JOSS in my opinion, and it is written well, especially by providing enough theoretical background of the problem. However, I noticed it lacks certain essential points:

* The installation procedure is inadequate for normal users. Users are referred to the official FEniCS installation guide for installing the prerequisites, which is a very advanced text for the users that `simcardems` is trying to target I guess. Having various versions of FEniCS makes it even more complicated since the software needs the legacy version. I strongly recommend that you revise the installation documentation by adding more explanation to it for normal users. Running FEniCS in Windows can be cumbersome, and I think it should be discussed in the installation procedure too. I tested the software using WSL (with Ubuntu 18.04) too and I should say that I had some problems, so the developers should explicitly mention if they support Windows or not, especially for running the Streamlit UI under WSL.
* The documentation is not consistent from the perspective of the Python command. You can find both `python` and `python3` in the procedures.
* The GUI is poorly documented. The same problem exists for the runtime flags and the arguments and parameters in the Python APIs. I suggest you add a page to the documentation to describe these parameters. The API documentation doesn’t help that much in my opinion since it provides the default value of the parameters only, which is also too technical for a normal user to go through. It’s not clear how the user is supposed to use the BCs on the GUI.
* The example factor files and ICs are never discussed in the documentation, and their usage is unclear from a practical point of view.

P.S. This issue is related to https://github.com/openjournals/joss-reviews/issues/4753
