### 2022-10-07 #4753 @mbarzegary
dear @AoifeHughes,

I confirm the quality of the software and the effort the authors have put in to develop it. Moreover, I know about the SimCardioTest project and its goals, so what the software tries to target as part of the project endeavor is clear to me. However, I believe there are some concerns that should be addressed before accepting the submission, elaborated in a couple of issues I have opened on the software repo (listed above). In my opinion, these are crucial to be fixed, allowing less-technical users to grab the code and use it (especially when the authors have pointed out that the code can be used for learning cardiac electro-mechanics). Advanced users can always enjoy and take advantage of the discretization schemes and model coupling in their codes.

I tried to be expressive since I know the value of the submitted work. I hope the authors find the comments useful.
