### 2022-10-07 a3ed8cb
Use buildx to build docker images so that we can build images for mulitple platforms - current solution only works for amd64
