### 2022-10-07 7c66c4f
Fix typo in docker-image workflow
