+++ .github/workflows/docker-image.yml
@@ -47,7 +47,7 @@ jobs:
           images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
 
       - name: Build and push Docker image
-        uses: docker/build-push-action@@v3
+        uses: docker/build-push-action@v3
         with:
           context: .
           push: true
