+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 2022.3.1
+current_version = 2022.3.2
 commit = True
 tag = True
 
