### 2022-10-07 8da90c8
Add latest tag to the most recent docker image
