+++ .github/workflows/docker-image.yml
@@ -52,5 +52,5 @@ jobs:
           context: .
           push: true
           platforms: linux/amd64,linux/arm64
-          tags: ${{ steps.meta.outputs.tags }}
+          tags: ${{ steps.meta.outputs.tags }} , latest
           labels: ${{ steps.meta.outputs.labels }}
