+++ docs/_toc.yml
@@ -4,6 +4,9 @@ parts:
   - caption: Installation
     chapters:
     - file: "install"
+    - file: "install_docker"
+    - file: "install_pip"
+    - file: "development_install"
   - caption: Mathematical Theory
     chapters:
     - file: "mechanics_model"
