### 2022-10-08 441ccbf
Build docs on each push but only deploy when pushing to master
