+++ .github/workflows/github-pages.yml
@@ -1,9 +1,6 @@
 name: github pages
 
-on:
-  push:
-    branches:
-      - master
+on: [push]
 
 jobs:
   deploy:
@@ -28,6 +25,7 @@ jobs:
 
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
+        if: github.ref == 'refs/heads/master'
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: ./docs/_build/html
