### 2022-10-08 1dd3b01
Swap out plotting in benchmark with plotly
