+++ docs/_config.yml
@@ -32,6 +32,8 @@ parse:
 sphinx:
   config:
     bibtex_bibfiles: ["refs.bib"]
+    html_js_files:
+      - https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js
 
   extra_extensions:
     - "sphinx.ext.autodoc"
