+++ docs/gui.md
@@ -15,7 +15,7 @@ Please also consult [Run docker](docker.md) for more options on how to use Docke
 
 In order to run the gui you need to first install `streamlit` and `fenics_plotly`, which can be done by either installing these packages separately, i.e
 ```
-python -m pip install streamlit fenics_plotly
+python3 -m pip install streamlit fenics_plotly
 ```
 or by installing `simcardems` with the extra requirements
 ```
@@ -26,7 +26,7 @@ python3 -m pip install "simcardems[gui]"
 
 You can run the gui from the command line by executing the following command
 ```
-python -m simcardems gui
+python3 -m simcardems gui
 ```
 This will open up a browser with an about page, and you can click on the `Simulation` radio button to get options about running a simulation. If no browser opens automatically, you can try to manually open the browser and go to the url http://localhost:8501
 
