### 2022-10-12 330eeb9
Remove hpc option - replaced by checking size of global communicator to turn of progress bar if needed
