+++ src/simcardems/config.py
@@ -20,7 +20,6 @@ class Config:
     )
     load_state: bool = False
     cell_init_file: utils.PathLike = ""
-    hpc: bool = False
     lx: float = 2.0
     ly: float = 0.7
     lz: float = 0.3
