### 2022-10-12 776334b
Replace hpc option by show_progress_bar/hide_progress_bar (more explicit)
