+++ src/simcardems/setup_models.py
@@ -498,6 +498,7 @@ class Runner:
         self,
         T: float = Config.T,
         save_freq: int = Config.save_freq,
+        show_progress_bar: bool = Config.show_progress_bar,
         st_progress: typing.Any = None,
     ):
         if not hasattr(self, "_outdir"):
@@ -511,7 +512,10 @@ class Runner:
 
         save_it = int(save_freq / self._dt)
         self.create_time_stepper(T, use_ns=True, st_progress=st_progress)
-        pbar = create_progressbar(time_stepper=self._time_stepper)
+        pbar = create_progressbar(
+            time_stepper=self._time_stepper,
+            show_progress_bar=show_progress_bar,
+        )
 
         # Store initial state
         save_state_every_n_beat = 5  # Save state every fifth beat
@@ -712,11 +716,12 @@ class TimeStepper:
 
 def create_progressbar(
     time_stepper: TimeStepper,
+    show_progress_bar: bool = Config.show_progress_bar,
 ):
-    hpc = bool(dolfin.MPI.size(dolfin.MPI.comm_world) > 1)
-    if hpc:
-        # Turn off progressbar
-        pbar = _tqdm(time_stepper, total=time_stepper.total_steps)
-    else:
+    if show_progress_bar:
+        # Show progressbar
         pbar = tqdm(time_stepper, total=time_stepper.total_steps)
+    else:
+        # Hide progressbar
+        pbar = _tqdm(time_stepper, total=time_stepper.total_steps)
     return pbar
