+++ _sources/cli.md
@@ -25,32 +25,47 @@ $ python3 -m simcardems run --help
 Usage: python -m simcardems run [OPTIONS]
 
 Options:
-  -o, --outdir PATH              Output directory
-  --dt FLOAT                     Time step
-  -T, --end-time FLOAT           End-time of simulation
-  -n, --num_refinements INTEGER  Number of refinements of for the mesh using
-                                 in the EP model
-  --save_freq INTEGER            Set frequency of saving results to file
-  --set_material TEXT            Choose material properties for mechanics
-                                 model (default is HolzapfelOgden, option is
-                                 Guccione
-  -dx FLOAT                      Spatial discretization
-  -lx FLOAT                      Size of mesh in x-direction
-  -ly FLOAT                      Size of mesh in y-direction
-  -lz FLOAT                      Size of mesh in z-direction
-  --bnd_cond [dirichlet|rigid]   Boundary conditions for the mechanics problem
-  --load_state                   If load existing state if exists, otherwise
-                                 create a new state
-  -IC, --cell_init_file TEXT     Path to file containing initial conditions
-                                 (json or h5 file). If none is provided then
-                                 the default initial conditions will be used
-  --loglevel INTEGER             How much printing. DEBUG: 10, INFO:20
-                                 (default), WARNING: 30
-  --hpc                          Indicate if simulations runs on hpc. This
-                                 turns off the progress bar.
-  --drug_factors_file TEXT       Set drugs scaling factors (json file)
-  --popu_factors_file TEXT       Set population scaling factors (json file)
-  --help                         Show this message and exit.
+  -o, --outdir PATH               Output directory
+  --dt FLOAT                      Time step
+  -T, --end-time FLOAT            End-time of simulation
+  -n, --num_refinements INTEGER   Number of refinements of for the mesh using
+                                  in the EP model
+  --save_freq INTEGER             Set frequency of saving results to file
+  --set_material TEXT             Choose material properties for mechanics
+                                  model (default is HolzapfelOgden, option is
+                                  Guccione
+  -dx FLOAT                       Spatial discretization
+  -lx FLOAT                       Size of mesh in x-direction
+  -ly FLOAT                       Size of mesh in y-direction
+  -lz FLOAT                       Size of mesh in z-direction
+  --bnd_cond [dirichlet|rigid]    Boundary conditions for the mechanics
+                                  problem
+  --load_state                    If load existing state if exists, otherwise
+                                  create a new state
+  -IC, --cell_init_file TEXT      Path to file containing initial conditions
+                                  (json or h5 file). If none is provided then
+                                  the default initial conditions will be used
+  --loglevel INTEGER              How much printing. DEBUG: 10, INFO:20
+                                  (default), WARNING: 30
+  --show_progress_bar / --hide_progress_bar
+                                  Shows or hide the progress bar.
+  --drug_factors_file TEXT        Set drugs scaling factors (json file)
+  --popu_factors_file TEXT        Set population scaling factors (json file)
+  --disease_state TEXT            Indicate disease state. Default is healthy.
+  --mechanics-ode-scheme [fd|bd|analytic]
+                                  Scheme used to solve the ODEs in the
+                                  mechanics model
+  --mechanics-use-continuation BOOLEAN
+                                  Use continuation based mechanics solver
+  --mechanics-use-custom-newton-solver BOOLEAN
+                                  Use custom newton solver and solve ODEs at
+                                  each Newton iteration
+  --pcl FLOAT                     Pacing cycle length (ms)
+  --spring FLOAT                  Set value of spring for Robin boundary
+                                  condition
+  --traction FLOAT                Set value of traction for Neumann boundary
+                                  condition
[…]
