### 2022-10-14 ComputationalPhysiology/simcardems#87 @cdaversin
Thanks for your comment.

Indeed, the (former) `hpc` option was only meant to turn on and off the progress bar, since the progress bar display is not always properly handled when running on HPC clusters. We agree that the name of this option was misleading, and this has been changed into `--show_progress_bar/--hide_progress_bar`.

Running simulations in parallel doesn't require any changes in the code, since FEniCS handles the parallel processing. We have updated the documentation regarding parallel simulations with the command that should be used.

We don't expect good scalability with the demo example due to the problem size, as you mention. But we agree that it would be good to have a larger problem illustrating the parallel computing capability.
