+++ docs/_toc.yml
@@ -6,7 +6,6 @@ parts:
     - file: "install"
     - file: "install_docker"
     - file: "install_pip"
-    - file: "development_install"
   - caption: Mathematical Theory
     chapters:
     - file: "mechanics_model"
