### 2022-10-16 #4753 @AoifeHughes
Hi, @mbarzegary and @finsberg - thanks for making progress on this!

Just a heads-up on a few things.

1. With regards to one review not being able to fully test the submission, I'll step in and finalise some tests myself and review the software which, depending on the other review should be fine. I've discussed this with other editors and that's all okay-ed.
2. I will be out of office until later in the week, I will be here all day Monday, so can respond to anything then.

Once again, thanks for all your hard work :)
