### 2022-10-16 ab8b7c5
Start decouling geometry and bc
