+++ src/simcardems/mechanics_model.py
@@ -1,6 +1,3 @@
-import typing
-from enum import Enum
-
 import dolfin
 import pulse
 from mpi4py import MPI
@@ -11,11 +8,6 @@ from .newton_solver import MechanicsNewtonSolver_ODE
 
 
 logger = utils.getLogger(__name__)
-
-
-class BoundaryConditions(str, Enum):
-    dirichlet = "dirichlet"
-    rigid = "rigid"
 
 
 class ContinuationBasedMechanicsProblem(pulse.MechanicsProblem):
@@ -63,8 +55,6 @@ class ContinuationBasedMechanicsProblem(pulse.MechanicsProblem):
 
 
 class MechanicsProblem(ContinuationBasedMechanicsProblem):
-    boundary_condition = BoundaryConditions.dirichlet
-
     def _init_spaces(self):
 
         mesh = self.geometry.mesh
@@ -139,8 +129,6 @@ class MechanicsProblem(ContinuationBasedMechanicsProblem):
 
 
 class RigidMotionProblem(MechanicsProblem):
-    boundary_condition = BoundaryConditions.rigid
-
     def _init_spaces(self):
 
         mesh = self.geometry.mesh
@@ -209,178 +197,3 @@ class RigidMotionProblem(MechanicsProblem):
         ]
 
         return sum(dolfin.dot(u, zi) * r[i] * dolfin.dx for i, zi in enumerate(RM))
-
-
-def setup_microstructure(mesh):
-    logger.debug("Set up microstructure")
-    V_f = dolfin.VectorFunctionSpace(mesh, "DG", 1)
-    f0 = dolfin.interpolate(
-        dolfin.Expression(("1.0", "0.0", "0.0"), degree=1, cell=mesh.ufl_cell()),
-        V_f,
-    )
-    s0 = dolfin.interpolate(
-        dolfin.Expression(("0.0", "1.0", "0.0"), degree=1, cell=mesh.ufl_cell()),
-        V_f,
-    )
-    n0 = dolfin.interpolate(
-        dolfin.Expression(("0.0", "0.0", "1.0"), degree=1, cell=mesh.ufl_cell()),
-        V_f,
-    )
-    # Collect the microstructure
-    return pulse.Microstructure(f0=f0, s0=s0, n0=n0)
-
-
-def float_to_constant(x: typing.Union[dolfin.Constant, float]) -> dolfin.Constant:
-    """Convert float to a dolfin constant.
-    If value is already a constant, do nothing.
-
-    Parameters
-    ----------
-    x : typing.Union[dolfin.Constant, float]
-        The value to be converted
-
-    Returns
-    -------
-    dolfin.Constant
-        The same value, wrapped in a constant
-    """
-    if isinstance(x, float):
-        return dolfin.Constant(x)
-    return x
-
-
-def setup_diriclet_bc(
-    mesh: dolfin.Mesh,
-    pre_stretch: typing.Optional[typing.Union[dolfin.Constant, float]] = None,
-    traction: typing.Union[dolfin.Constant, float] = None,
-    spring: typing.Union[dolfin.Constant, float] = None,
-    fix_right_plane: bool = False,
-) -> typing.Tuple[pulse.BoundaryConditions, pulse.MarkerFunctions]:
-    """Completely fix the left side of the mesh in the x-direction (i.e the side with the
-    lowest x-values), fix the plane with y=0 in the y-direction, fix the plane with
-    z=0 in the z-direction and apply some boundary condition to the right side.
-
-
-    Parameters
-    ----------
-    mesh : dolfin.Mesh
-        A cube or box-shaped mesh
-    pre_stretch : typing.Union[dolfin.Constant, float], optional
-        Value representing the amount of pre stretch, by default None
-    traction : typing.Union[dolfin.Constant, float], optional
[…]
