+++ src/simcardems/geometry.py
@@ -1,13 +1,31 @@
 import abc
-from typing import Dict
+import json
+from typing import Any, Dict
 from typing import Optional
+from pathlib import Path
+import json
 
 import dolfin
 import numpy as np
+import pulse
+
+from simcardems.config import Config
+
 
 from . import utils
 
 logger = utils.getLogger(__name__)
+
+
+def load_geometry(folder: str) -> "BaseGeometry":
+    info_path = Path(folder) / "info.json"
+    if not info_path.is_file():
+        raise RuntimeError("Cannot find info.json in geometry folder")
+
+    info = json.loads(info_path.read_text())
+    mesh_type = info.get("mesh_type")
+    if mesh_type is None:
+        raise RuntimeError(f"{info_path} missing key 'mesh_type'")
 
 
 def refine_mesh(
@@ -23,18 +41,18 @@ def refine_mesh(
     return mesh
 
 
-def create_boxmesh(Lx, Ly, Lz, dx=0.5, refinements=0):
-    # Create computational domain [0, Lx] x [0, Ly] x [0, Lz]
+def create_boxmesh(lx, ly, lz, dx=0.5, refinements=0):
+    # Create computational domain [0, lx] x [0, ly] x [0, lz]
     # with resolution prescribed by benchmark or more refinements
 
     N = lambda v: int(np.rint(v))
     mesh = dolfin.BoxMesh(
         dolfin.MPI.comm_world,
         dolfin.Point(0.0, 0.0, 0.0),
-        dolfin.Point(Lx, Ly, Lz),
-        N(Lx / dx),
-        N(Ly / dx),
-        N(Lz / dx),
+        dolfin.Point(lx, ly, lz),
+        N(lx / dx),
+        N(ly / dx),
+        N(lz / dx),
     )
 
     for i in range(refinements):
@@ -49,6 +67,9 @@ class BaseGeometry(abc.ABC):
 
     num_refinements: int = 0
     mechanics_mesh: dolfin.Mesh
+    ffun: dolfin.MeshFunction
+    markers: Dict[str, int]
+    microstructure: pulse.Microstructure
 
     @property
     def ep_mesh(self) -> dolfin.Mesh:
@@ -60,11 +81,37 @@ class BaseGeometry(abc.ABC):
     def parameters(self) -> Dict[str, float]:
         ...
 
+    @property
+    def marker_functions(self):
+        return pulse.MarkerFunctions(ffun=self.ffun)
+
+    @property
+    def f0(self) -> dolfin.Function:
+        return self.microstructure.f0
+
+    @property
+    def s0(self) -> dolfin.Function:
+        return self.microstructure.s0
+
+    @property
+    def n0(self) -> dolfin.Function:
+        return self.microstructure.n0
+
 
 class Geometry(BaseGeometry):
-    def __init__(self, mesh: dolfin.Mesh, num_refinements: int = 0) -> None:
+    def __init__(
+        self,
+        mesh: dolfin.Mesh,
+        ffun: dolfin.MeshFunction,
+        microstructure: pulse.Microstructure,
+        markers: Optional[Dict[str, int]] = None,
+        num_refinements: int = 0,
+    ) -> None:
         self.num_refinements = num_refinements
         self.mechanics_mesh = mesh
+        self.ffun = ffun
+        self.markers = markers
+        self.microstructure = microstructure
 
     def __repr__(self):
         return (
@@ -78,24 +125,34 @@ class Geometry(BaseGeometry):
 
 
 class SlabGeometry(BaseGeometry):
+
+    markers: Dict[str, int] = {"left": 1, "right": 2, "plane_y0": 3, "plane_z0": 4}
+
     def __init__(
         self,
-        lx: float,
[…]
