### 2022-10-17 0201d61
Implement a few more properties on the geometry class
