### 2022-10-18 ComputationalPhysiology/simcardems#85 @IlsevanHerck
Thank you for the comments @mbarzegary.

I just have 2 short questions regarding the ParaView visualisation:
1. How did you open the xdmf-files? Via the 'load data' option in ParaView, open ParaView from the xdmf file, or another way?
2. Does the xdmf file have a corresponding .h5 file in the same folder?
