### 2022-10-18 ComputationalPhysiology/simcardems#85 @mbarzegary
hi,
1) I open them in PV the way I do for fenics results, like by "File->Open" or by dropping the the xdmf file on the pipeline browser.
2) yes, the files have corresponding h5 files beside. I share the results I got for demo example for your reference if it helps: https://we.tl/t-ZrxRmMlJsB. the link expires in 7 days.
