+++ src/simcardems/geometry.py
@@ -5,6 +5,7 @@ from typing import Any
 from typing import Dict
 from typing import Optional
 from typing import Tuple
+from typing import Union
 
 import dolfin
 import numpy as np
@@ -71,16 +72,11 @@ def create_boxmesh(lx, ly, lz, dx=0.5, refinements=0):
 class BaseGeometry(abc.ABC):
     """Abstract geometry base class"""
 
-    num_refinements: int = 0
     mechanics_mesh: dolfin.Mesh
     ffun: dolfin.MeshFunction
     markers: Dict[str, Tuple[int, int]]
     microstructure: pulse.Microstructure
-
-    @property
-    @abc.abstractmethod
-    def parameters(self) -> Dict[str, Any]:
-        ...
+    parameters: Dict[str, Any]
 
     @property
     def mesh(self) -> dolfin.Mesh:
@@ -99,13 +95,30 @@ class BaseGeometry(abc.ABC):
         return dolfin.ds(domain=self.mesh, subdomain_data=self.ffun)
 
     @property
+    def num_refinements(self) -> int:
+        return self.parameters["num_refinements"]
+
+    def dump(self, fname: utils.PathLike, file_mode="a"):
+        path = Path(fname)
+        with dolfin.HDF5File(
+            self.mesh.mpi_comm(),
+            path.as_posix(),
+            file_mode,
+        ) as h5file:
+            h5file.write(self.ep_mesh, "geometry/ep_mesh")
+            h5file.write(self.mechanics_mesh, "geometry/mechanics_mesh")
+
+    @property
     def ep_mesh(self) -> dolfin.Mesh:
         return self._ep_mesh
 
     @ep_mesh.setter
     def ep_mesh(self, mesh: Optional[dolfin.Mesh]) -> None:
         if mesh is None:
-            self._ep_mesh = refine_mesh(self.mechanics_mesh, self.num_refinements)
+            self._ep_mesh = refine_mesh(
+                self.mechanics_mesh,
+                self.parameters["num_refinements"],
+            )
             if self.outdir is not None:
                 mesh_path = self.outdir / "ep_mesh.xdmf"
                 with dolfin.XDMFFile(mesh_path.as_posix()) as f:
@@ -149,38 +162,32 @@ class Geometry(BaseGeometry):
         ffun: dolfin.MeshFunction,
         microstructure: pulse.Microstructure,
         markers: Optional[Dict[str, Tuple[int, int]]] = None,
-        num_refinements: int = 0,
+        parameters: Optional[Dict[str, Union[int, str]]] = None,
     ) -> None:
-        self.num_refinements = num_refinements
         self.mechanics_mesh = mesh
         self.ffun = ffun
         self.markers = markers or {}
         self.microstructure = microstructure
+        self.parameters = Geometry.default_parameters()
+        if parameters is not None:
+            self.parameters.update(parameters)
 
     def __repr__(self):
         return (
             f"{self.__class__.__name__}("
             f"mesh={self.mechanics_mesh}, "
-            f"num_refinements={self.num_refinements})"
+            f"num_refinements={self.parameters['num_refinements']})"
         )
 
-    def parameters(self):
-        return {"num_refinements": self.num_refinements}
+    @staticmethod
+    def default_parameters():
+        return {"num_refinements": 1, "fiber_space": "DG_1"}
 
 
 class SlabGeometry(BaseGeometry):
-
-    markers: Dict[str, Tuple[int, int]] = {
-        "X0": (2, 1),
-        "X1'": (2, 2),
-        "Y0": (2, 3),
-        "Z0": (2, 4),
-    }
-
     def __init__(
         self,
         parameters: Optional[Dict[str, Any]] = None,
[…]
