### 2022-10-24 ac6d881
Make is possible to pass microstructure to ep model
