+++ tests/test_mechanics_model.py
@@ -58,7 +58,6 @@ def test_boundary_conditions(
     assert np.isclose(U(1.0, 0.5, 0.5), expected_u, atol=1e-5).all()
 
 
-@pytest.mark.xfail
 @pytest.mark.parametrize(
     "pre_stretch, traction, spring, fix_right_plane, expected_u",
     [
