+++ setup.cfg
@@ -55,6 +55,8 @@ dev =
 docs =
     jupyter-book
     jupytext
+    pandas
+    plotly
     sphinxcontrib-bibtex
 gui =
     fenics-plotly
