+++ src/simcardems/geometry.py
@@ -5,29 +5,31 @@ from typing import Any
 from typing import Dict
 from typing import Optional
 from typing import Tuple
-from typing import Union
 
 import dolfin
 import numpy as np
 import pulse
+from cardiac_geometries.geometry import Geometry
+from cardiac_geometries.geometry import H5Path
 
 from . import utils
 
 logger = utils.getLogger(__name__)
 
 
-def load_geometry(folder: utils.PathLike) -> "BaseGeometry":
-    info_path = Path(folder) / "info.json"
-    if not info_path.is_file():
-        raise RuntimeError("Cannot find info.json in geometry folder")
+def load_geometry(mesh_path: utils.PathLike, schema_path: utils.PathLike):
+    geo = Geometry.from_file(fname=mesh_path, schema_path=schema_path)
 
-    info = json.loads(info_path.read_text())
+    info = getattr(geo, "info", None)
+    if info is None:
+        raise RuntimeError("Unable to load info from geometry")
+
     mesh_type = info.get("mesh_type")
     if mesh_type is None:
-        raise RuntimeError(f"{info_path} missing key 'mesh_type'")
+        raise RuntimeError("Unable to get mesh type from info")
 
     if mesh_type == "slab":
-        return SlabGeometry.from_folder(folder)
+        return SlabGeometry.from_geometry(geo)
     elif mesh_type == "lv_ellipsoid":
         raise NotImplementedError
     elif mesh_type == "biv_ellipsoid":
@@ -72,11 +74,126 @@ def create_boxmesh(lx, ly, lz, dx=0.5, refinements=0):
 class BaseGeometry(abc.ABC):
     """Abstract geometry base class"""
 
-    mechanics_mesh: dolfin.Mesh
-    ffun: dolfin.MeshFunction
-    markers: Dict[str, Tuple[int, int]]
-    microstructure: pulse.Microstructure
-    parameters: Dict[str, Any]
+    def __init__(
+        self,
+        parameters: Optional[Dict[str, Any]] = None,
+        mechanics_mesh: Optional[dolfin.Mesh] = None,
+        ep_mesh: Optional[dolfin.Mesh] = None,
+        microstructure: Optional[pulse.Microstructure] = None,
+        microstructure_ep: Optional[pulse.Microstructure] = None,
+        ffun: Optional[dolfin.MeshFunction] = None,
+        markers: Optional[Dict[str, Tuple[int, int]]] = None,
+        outdir: Optional[utils.PathLike] = None,
+    ) -> None:
+
+        self.markers = type(self).default_markers()
+        if markers is not None:
+            self.markers.update(markers)
+
+        self.outdir = outdir  # type: ignore
+        self.parameters = type(self).default_parameters()
+        if parameters is not None:
+            self.parameters.update(parameters)
+
+        self.mechanics_mesh = mechanics_mesh
+        self.ep_mesh = ep_mesh
+
+        self.ffun = ffun
+        self.microstructure = microstructure
+        self.microstructure_ep = microstructure_ep
+        self.validate()
+
+    def __eq__(self, __o: object) -> bool:
+        if not isinstance(__o, type(self)):
+            return NotImplemented
+        # TODO: We might add more checks here
+        return self.parameters == __o.parameters
+
+    def validate(self):
+        pass
+
+    @staticmethod
+    @abc.abstractmethod
+    def default_markers() -> Dict[str, Tuple[int, int]]:
+        ...
+
+    @staticmethod
+    @abc.abstractmethod
+    def default_parameters() -> Dict[str, Any]:
+        ...
+
+    @staticmethod
+    def default_schema() -> Dict[str, H5Path]:
+        return {
+            "mesh": H5Path(
+                h5group="/geometry/mesh/mechanics",
+                is_mesh=True,
+            ),
[…]
