+++ setup.cfg
@@ -18,6 +18,7 @@ keywords = action potential, cardiac mechanics, electrophysiology, electromechan
 packages = find:
 install_requires =
     ap-features
+    cardiac-geometries
     cbcbeat
     click
     fenics-pulse

[…]
