### 2022-11-01 8e9ffac
Refactor boundary conditions
