+++ src/simcardems/setup_models.py
@@ -1,14 +1,13 @@
-import logging
+import math
 import typing
-from collections import namedtuple
 from pathlib import Path
+from typing import NamedTuple
 
 import cbcbeat
 import dolfin
 import pulse
 from tqdm import tqdm
 
-from . import boundary_conditions
 from . import config
 from . import em_model
 from . import ep_model
@@ -17,14 +16,18 @@ from . import land_model
 from . import mechanics_model
 from . import save_load_functions as io
 from . import utils
+from .em_model import EMCoupling
 from .ORdmm_Land import ORdmm_Land as CellModel
 
 logger = utils.getLogger(__name__)
 
-EMState = namedtuple(
-    "EMState",
-    ["coupling", "solver", "mech_heart", "t0"],
-)
+
+class EMState(NamedTuple):
+    coupling: EMCoupling
+    solver: cbcbeat.SplittingSolver
+    mech_heart: pulse.MechanicsProblem
+    geometry: geometry.BaseGeometry
+    t0: float
 
 
 def setup_EM_model(config: config.Config):
@@ -67,28 +70,9 @@ def setup_EM_model(config: config.Config):
         coupling=coupling,
         solver=solver,
         mech_heart=mech_heart,
+        geometry=geo,
         t0=0,
     )
-
-
-def resolve_boundary_conditions(
-    geo: geometry.BaseGeometry,
-    pre_stretch: typing.Optional[typing.Union[dolfin.Constant, float]] = None,
-    traction: typing.Union[dolfin.Constant, float] = None,
-    spring: typing.Union[dolfin.Constant, float] = None,
-    fix_right_plane: bool = config.Config.fix_right_plane,
-) -> boundary_conditions.BaseBoundaryConditions:
-    if isinstance(geo, geometry.SlabGeometry):
-        return boundary_conditions.SlabBoundaryConditions(
-            geo=geo,
-            pre_stretch=pre_stretch,
-            traction=traction,
-            spring=spring,
-            fix_right_plane=fix_right_plane,
-        )
-    else:
-        # TODO: Implement more boundary conditions
-        raise NotImplementedError
 
 
 def setup_mechanics_solver(
@@ -111,14 +95,6 @@ def setup_mechanics_solver(
 ):
     """Setup mechanics model with dirichlet boundary conditions or rigid motion."""
     logger.info("Set up mechanics model")
-
-    bcs = resolve_boundary_conditions(
-        geo=geo,
-        pre_stretch=pre_stretch,
-        traction=traction,
-        spring=spring,
-        fix_right_plane=fix_right_plane,
-    )
 
     # Use parameters from Biaxial test in Holzapfel 2019 (Table 1).
     material_parameters = dict(
@@ -164,16 +140,15 @@ def setup_mechanics_solver(
             active_model=active_model,
         )
 
-    Problem = mechanics_model.MechanicsProblem
-    if bnd_cond == config.SlabBoundaryConditionTypes.rigid:
-        Problem = mechanics_model.RigidMotionProblem
-
-    verbose = logger.getEffectiveLevel() < logging.INFO
-    problem = Problem(
-        geo,
-        material,
-        bcs(),
-        solver_parameters={"linear_solver": linear_solver, "verbose": verbose},
+    problem = mechanics_model.create_slab_problem(
+        material=material,
+        geo=geo,
+        bnd_cond=bnd_cond,
+        pre_stretch=pre_stretch,
+        traction=traction,
+        spring=spring,
+        fix_right_plane=fix_right_plane,
+        linear_solver=linear_solver,
         use_custom_newton_solver=use_custom_newton_solver,
     )
 
@@ -268,7 +243,7 @@ class Runner:
         if not reset and self._state_path.is_file():
[…]
