### 2022-11-01 042f7d1
Fix command line interface
