+++ tests/test_boundary_conditions.py
@@ -3,16 +3,14 @@ import math
 import simcardems
 
 
-def test_SlabBoundaryConditions():
-    geo = simcardems.geometry.SlabGeometry(parameters=dict(lx=1, ly=1, lz=1, dx=1))
-    bc = simcardems.boundary_conditions.SlabBoundaryConditions(
+def test_SlabBoundaryConditions(geo):
+    bcs = simcardems.boundary_conditions.create_slab_boundary_conditions(
         geo=geo,
         pre_stretch=0.1,
         traction=1.0,
         spring=0.2,
         fix_right_plane=False,
     )
-    bcs = bc()
 
     assert len(bcs.neumann) == 1
     assert math.isclose(float(bcs.neumann[0].traction), 1.0)
