### 2022-11-01 e5beb48
Add missing slab geometry
