+++ src/simcardems/postprocess.py
@@ -1,262 +1,17 @@
 import json
-import warnings
-from enum import Enum
 from pathlib import Path
 
-import ap_features as apf
 import dolfin
-import h5py
 import matplotlib.pyplot as plt
 import numpy as np
 import tqdm
 
+import ap_features as apf
 from . import utils
 from .datacollector import DataGroups
 from .datacollector import DataLoader
 
 logger = utils.getLogger(__name__)
-
-
-def center_func(fmin, fmax):
-    return fmin + (fmax - fmin) / 2
-
-
-class BoundaryNodes(Enum):
-    center = "center"
-    node_A = "node_A"
-    node_B = "node_B"
-    node_C = "node_C"
-    node_D = "node_D"
-    node_E = "node_E"
-    node_F = "node_F"
-    node_G = "node_G"
-    node_H = "node_H"
-    xmax = "xmax"
-    xmin = "xmin"
-    ymax = "ymax"
-    ymin = "ymin"
-    zmax = "zmax"
-    zmin = "zmin"
-
-
-class Boundary:
-    def __init__(self, mesh):
-        self.mesh = mesh
-
-    @staticmethod
-    def nodes():
-        return BoundaryNodes._member_names_
-
-    @property
-    def boundaries(self):
-        coords = self.mesh.coordinates()
-        return dict(
-            max_x=coords.T[0].max(),
-            min_x=coords.T[0].min(),
-            max_y=coords.T[1].max(),
-            min_y=coords.T[1].min(),
-            max_z=coords.T[2].max(),
-            min_z=coords.T[2].min(),
-        )
-
-    @property
-    def xmin(self):
-        return [
-            self.boundaries["min_x"],
-            center_func(self.boundaries["min_y"], self.boundaries["max_y"]),
-            center_func(self.boundaries["min_z"], self.boundaries["max_z"]),
-        ]
-
-    @property
-    def xmax(self):
-        return [
-            self.boundaries["max_x"],
-            center_func(self.boundaries["min_y"], self.boundaries["max_y"]),
-            center_func(self.boundaries["min_z"], self.boundaries["max_z"]),
-        ]
-
-    @property
-    def ymin(self):
-        return [
-            center_func(self.boundaries["min_x"], self.boundaries["max_x"]),
-            self.boundaries["min_y"],
-            center_func(self.boundaries["min_z"], self.boundaries["max_z"]),
-        ]
-
-    @property
-    def ymax(self):
-        return [
-            center_func(self.boundaries["min_x"], self.boundaries["max_x"]),
-            self.boundaries["max_y"],
-            center_func(self.boundaries["min_z"], self.boundaries["max_z"]),
-        ]
-
-    @property
-    def zmin(self):
-        return [
-            center_func(self.boundaries["min_x"], self.boundaries["max_x"]),
-            center_func(self.boundaries["min_y"], self.boundaries["max_y"]),
-            self.boundaries["min_z"],
-        ]
-
-    @property
-    def zmax(self):
-        return [
-            center_func(self.boundaries["min_x"], self.boundaries["max_x"]),
-            center_func(self.boundaries["min_y"], self.boundaries["max_y"]),
-            self.boundaries["max_z"],
-        ]
-
-    @property
-    def center(self):
-        return [
-            center_func(self.boundaries["min_x"], self.boundaries["max_x"]),
-            center_func(self.boundaries["min_y"], self.boundaries["max_y"]),
-            center_func(self.boundaries["min_z"], self.boundaries["max_z"]),
-        ]
-
-    @property
-    def node_A(self):
-        return [
-            self.boundaries["min_x"],
-            self.boundaries["min_y"],
-            self.boundaries["min_z"],
-        ]
-
-    @property
-    def node_B(self):
-        return [
-            self.boundaries["max_x"],
[…]
