### 2022-11-02 a25c4a5
Fix tests for datacollector
