+++ tests/test_datacollector.py
@@ -8,34 +8,34 @@ import pytest
 import simcardems
 
 
-def test_DataCollector_reset_state_when_file_exists(mesh):
+def test_DataCollector_reset_state_when_file_exists(geo):
     outdir = Path("testdir")
 
-    simcardems.DataCollector(outdir, mesh, mesh)
+    simcardems.DataCollector(outdir, geo=geo)
 
     with mock.patch("simcardems.utils.remove_file") as remove_file_mock:
-        collector = simcardems.DataCollector(outdir, mesh, mesh, reset_state=True)
+        collector = simcardems.DataCollector(outdir, geo=geo, reset_state=True)
     remove_file_mock.assert_called()
     assert Path(collector.results_file).is_file()
     shutil.rmtree(outdir)
 
 
-def test_DataCollector_not_reset_state_when_file_exists(mesh):
+def test_DataCollector_not_reset_state_when_file_exists(geo):
     outdir = Path("testdir")
 
-    simcardems.DataCollector(outdir, mesh, mesh)
+    simcardems.DataCollector(outdir, geo=geo)
 
     with mock.patch("simcardems.utils.remove_file") as remove_file_mock:
-        collector = simcardems.DataCollector(outdir, mesh, mesh, reset_state=False)
+        collector = simcardems.DataCollector(outdir, geo=geo, reset_state=False)
     remove_file_mock.assert_not_called()
     assert Path(collector.results_file).is_file()
     shutil.rmtree(outdir)
 
 
-def test_DataCollector_create_file_with_mesh(mesh):
+def test_DataCollector_create_file_with_mesh(geo):
     outdir = Path("testdir")
 
-    collector = simcardems.DataCollector(outdir, mesh, mesh)
+    collector = simcardems.DataCollector(outdir, geo=geo)
     assert Path(collector.results_file).is_file()
     with h5py.File(collector.results_file, "r") as h5file:
         assert "ep/mesh" in h5file
@@ -43,7 +43,7 @@ def test_DataCollector_create_file_with_mesh(mesh):
     shutil.rmtree(outdir)
 
 
-def test_DataCollector_register(mesh):
+def test_DataCollector_register(geo):
     V = dolfin.FunctionSpace(mesh, "CG", 1)
     f = dolfin.Function(V)
     outdir = Path("testdir")
@@ -79,10 +79,10 @@ def test_DataCollector_store(group, mesh):
     shutil.rmtree(outdir)
 
 
-def test_DataLoader_load_empty_files_raises_ValueError(mesh):
+def test_DataLoader_load_empty_files_raises_ValueError(geo):
     outdir = Path("testdir")
 
-    collector = simcardems.DataCollector(outdir, mesh, mesh)
+    collector = simcardems.DataCollector(outdir, geo=geo)
     with pytest.raises(ValueError):
         simcardems.DataLoader(collector.results_file)
     shutil.rmtree(outdir)
