### 2022-11-02 dbbaca5
Correct path to geometries in simple demo and release test (in demos-folder)
