### 2022-11-02 8e22483
Fix tests in datacollector
