+++ tests/test_datacollector.py
@@ -1,4 +1,3 @@
-import shutil
 from pathlib import Path
 from unittest import mock
 
@@ -8,81 +7,72 @@ import pytest
 import simcardems
 
 
-def test_DataCollector_reset_state_when_file_exists(geo):
-    outdir = Path("testdir")
+def test_DataCollector_reset_state_when_file_exists(tmp_path, geo):
 
-    simcardems.DataCollector(outdir, geo=geo)
+    simcardems.DataCollector(tmp_path, geo=geo)
 
     with mock.patch("simcardems.utils.remove_file") as remove_file_mock:
-        collector = simcardems.DataCollector(outdir, geo=geo, reset_state=True)
+        collector = simcardems.DataCollector(tmp_path, geo=geo, reset_state=True)
     remove_file_mock.assert_called()
     assert Path(collector.results_file).is_file()
-    shutil.rmtree(outdir)
 
 
-def test_DataCollector_not_reset_state_when_file_exists(geo):
-    outdir = Path("testdir")
+def test_DataCollector_not_reset_state_when_file_exists(tmp_path, geo):
 
-    simcardems.DataCollector(outdir, geo=geo)
+    simcardems.DataCollector(tmp_path, geo=geo)
 
     with mock.patch("simcardems.utils.remove_file") as remove_file_mock:
-        collector = simcardems.DataCollector(outdir, geo=geo, reset_state=False)
+        collector = simcardems.DataCollector(tmp_path, geo=geo, reset_state=False)
     remove_file_mock.assert_not_called()
     assert Path(collector.results_file).is_file()
-    shutil.rmtree(outdir)
 
 
-def test_DataCollector_create_file_with_mesh(geo):
-    outdir = Path("testdir")
+def test_DataCollector_create_file_with_geo(tmp_path, geo):
 
-    collector = simcardems.DataCollector(outdir, geo=geo)
+    collector = simcardems.DataCollector(tmp_path, geo=geo)
     assert Path(collector.results_file).is_file()
     with h5py.File(collector.results_file, "r") as h5file:
-        assert "ep/mesh" in h5file
-        assert "mechanics/mesh" in h5file
-    shutil.rmtree(outdir)
+        assert "geometry" in h5file
 
 
-def test_DataCollector_register(geo):
-    V = dolfin.FunctionSpace(mesh, "CG", 1)
+def test_DataCollector_register(tmp_path, geo):
+    V = dolfin.FunctionSpace(geo.mesh, "CG", 1)
     f = dolfin.Function(V)
-    outdir = Path("testdir")
+
     simcardems.set_log_level(10)
-    collector = simcardems.DataCollector(outdir, mech_mesh=mesh, ep_mesh=mesh)
+    collector = simcardems.DataCollector(tmp_path, geo=geo)
     collector.register("ep", "func", f)
     assert "func" in collector.names["ep"]
 
     with pytest.raises(ValueError) as ex:
         simcardems.DataLoader(collector.results_file)
     assert str(ex.value) == "No functions found in results file"
-    shutil.rmtree(outdir)
 
 
 @pytest.mark.parametrize("group", ["ep", "mechanics"])
-def test_DataCollector_store(group, mesh):
+def test_DataCollector_store(group, geo, tmp_path):
+    mesh = geo.mesh if group == "mechanics" else geo.ep_mesh
     V = dolfin.FunctionSpace(mesh, "CG", 1)
     f = dolfin.Function(V)
     f.vector()[:] = 42
-    outdir = Path("testdir")
+
     simcardems.set_log_level(10)
-    collector = simcardems.DataCollector(outdir, mech_mesh=mesh, ep_mesh=mesh)
+    collector = simcardems.DataCollector(tmp_path, geo=geo)
[…]
