### 2022-11-02 8124f71
Fix benchmark and update paths so that they are relative to the path of the demo file
