+++ demos/ventricular_geometry.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 # # Left ventricular geometry
 #
-# In this demo we will run the same
+# In this demo we will run the same simple demo but using a ventricular geometry instead.
+# Note that we currently only support a fixed base boundary condition, but plan to add support for
+# more complicated boundary conditions in the future.
 #
 # Import the necessary libraries
 #
@@ -44,7 +46,7 @@ config = simcardems.Config(
 #  'geometry_path': 'demos/geometries/slab.h5',
 #  'geometry_schema_path': 'demos/geometries/slab.json',
 #  'linear_mechanics_solver': 'mumps',
-#  'load_state': True,
+#  'load_state': False,
 #  'loglevel': 20,
 #  'mechanics_ode_scheme': <Scheme.analytic: 'analytic'>,
 #  'mechanics_use_continuation': False,
@@ -73,11 +75,14 @@ runner.solve(T=config.T, save_freq=config.save_freq, show_progress_bar=True)
 # ```
 # results_simple_demo
 # ├── results.h5
+# ├── results.json
 # ├── state.h5
+# ├── state.json
 # ```
 # The file `state.h5` contains the final state which can be used if you want use the final state as a starting point for the next simulation.
-# The file `results.h5` contains the Displacement ($u$), active tension ($T_a$), voltage ($V$) and calcium ($Ca$) for each time step.
-# We can also plot the traces using the postprocess module
+# The file `results.h5` contains different traces such as Displacement ($u$), active tension ($T_a$), voltage ($V$) and calcium ($Ca$) for each time step.
+# The `.json` files are used to load load the geometries and, see https://computationalphysiology.github.io/cardiac_geometries/quickstart.html#schema for more info.
+# We can also plot the traces using the postprocess module.
 #
 #
 
@@ -85,7 +90,20 @@ runner.solve(T=config.T, save_freq=config.save_freq, show_progress_bar=True)
 simcardems.postprocess.plot_state_traces(outdir.joinpath("results.h5"))
 
 #
-# And save the output to xdmf-files that can be viewed in Paraview
+# Here the traces generated by averaging over this domain.
+
+
+# This will create a figure in the output directory called `state_traces.png` which in this case is shown in {numref}`Figure {number} <lv_demo_state_traces>` we see the resulting state traces, and can also see the instant drop in the active tension ($T_a$) at the time of the triggered release.
+#
+# ```{figure} figures/lv_demo_state_traces.png
+# ---
+# name: lv_demo_state_traces
+# ---
+# Traces of the stretch ($\lambda$), the active tension ($T_a$), the membrane potential ($V$) and the intercellular calcium concentration ($Ca$) at the center of the geometry.
+# ```
+
+#
+# We can also save the output to xdmf-files that can be viewed in Paraview
 #
 
 simcardems.postprocess.make_xdmffiles(outdir.joinpath("results.h5"))
@@ -99,14 +117,4 @@ simcardems.postprocess.make_xdmffiles(outdir.joinpath("results.h5"))
 # ---
 #
 # Displacement ($u$), active tension ($T_a$), voltage ($V$) and calcium ($Ca$) visualized for a specific time point in Paraview.
-# ```
-
-
-# This will create a figure in the output directory called `state_traces.png` which in this case is shown in {numref}`Figure {number} <lv_demo_state_traces>` we see the resulting state traces, and can also see the instant drop in the active tension ($T_a$) at the time of the triggered release.
-#
-# ```{figure} figures/lv_demo_state_traces.png
-# ---
-# name: lv_demo_state_traces
-# ---
-# Traces of the stretch ($\lambda$), the active tension ($T_a$), the membrane potential ($V$) and the intercellular calcium concentration ($Ca$) at the center of the geometry.
 # ```
