### 2022-11-03 dad193e
Change bnd_cond to a flag (bnd_rigid) for using Rigid Motion conditions
