### 2022-11-03 c7cb3ef
Add bnd_rigid to cspell dict
