### 2022-11-03 94e6119
Fix test on em model
