+++ tests/test_em_model.py
@@ -5,8 +5,8 @@ import simcardems
 
 
 @pytest.mark.slow
-@pytest.mark.parametrize("bnd_cond", ["dirichlet", "rigid"])
-def test_em_model(coupling, ep_solver, cell_params, geo, bnd_cond):
+@pytest.mark.parametrize("bnd_rigid", [False, True])
+def test_em_model(coupling, ep_solver, cell_params, geo, bnd_rigid):
     coupling.register_ep_model(ep_solver)
     with mock.patch(
         "simcardems.setup_models.pulse.mechanicsproblem.MechanicsProblem.solve",
@@ -14,7 +14,7 @@ def test_em_model(coupling, ep_solver, cell_params, geo, bnd_cond):
         solve_mock.return_value = (1, True)  # (niter, nconv)
         simcardems.setup_models.setup_mechanics_solver(
             coupling=coupling,
-            bnd_cond=bnd_cond,
+            bnd_rigid=bnd_rigid,
             cell_params=cell_params,
             geo=geo,
         )
