### 2022-11-04 ComputationalPhysiology/simcardems#85 @jorgensd
It seems to be an issue with `boost` and string concatenation when generating the XDMFFile on your system.
This has been fixed in DOLFINx: https://github.com/FEniCS/dolfinx/pull/1743, https://github.com/FEniCS/dolfinx/pull/1826
but has not been patched in DOLFIN.
For instance: https://bitbucket.org/fenics-project/dolfin/src/74a6ac2095059ea6582fe4ea11686c1c723e14f5/dolfin/io/XDMFFile.cpp#lines-1822 would have to be changed.
