+++ setup.cfg
@@ -26,7 +26,7 @@ install_requires =
     numpy
     scipy
     tqdm
-python_requires = >=3.6
+python_requires = >=3.7
 package_dir =
     =src
 zip_safe = False
