### 2022-11-11 ComputationalPhysiology/simcardems#85 @jorgensd
Fix proposed in: https://bitbucket.org/fenics-project/dolfin/pull-requests/555/fix-xdmffile-string-merging-with-h5
