### 2022-11-11 c686a16
Use enum mesh type to avoid using strings in paralell
