### 2022-11-11 80aeb8d
Do not use official fenics image since it is using python3.6 which is no longer supported
