### 2022-11-11 ba7c8a3
Remove linting in CI - this is done by pre-commit
