+++ .github/workflows/main.yml
@@ -30,14 +30,6 @@ jobs:
           python3 -m pip install h5py --no-binary=h5py
           python3 -m pip install -e ".[dev]"
 
-      - name: Run flake8
-        run: |
-          python3 -m flake8 src tests
-
-      - name: Run black
-        run: |
-          python3 -m black --check src tests
-
       - name: Test with pytest
         run: |
           python3 -m pytest
