+++ .github/workflows/main.yml
@@ -46,7 +46,6 @@ jobs:
           python3 -m
           pip install
           build
-          --user
 
       - name: Build a binary wheel and a source tarball
         run: >-
