### 2022-11-12 65304bc
Adapt ffun and recreate microstructure if using qudrature spaces
