+++ src/simcardems/geometry.py
@@ -54,6 +54,7 @@ def refine_mesh(
     redistribute: bool = False,
 ) -> dolfin.Mesh:
 
+    dolfin.parameters["refinement_algorithm"] = "plaza_with_parent_facets"
     for i in range(num_refinements):
         logger.info(f"Performing refinement {i+1}")
         mesh = dolfin.refine(mesh, redistribute=redistribute)
@@ -93,6 +94,7 @@ class BaseGeometry(abc.ABC):
         microstructure: Optional[pulse.Microstructure] = None,
         microstructure_ep: Optional[pulse.Microstructure] = None,
         ffun: Optional[dolfin.MeshFunction] = None,
+        ffun_ep: Optional[dolfin.MeshFunction] = None,
         markers: Optional[Dict[str, Tuple[int, int]]] = None,
         outdir: Optional[utils.PathLike] = None,
     ) -> None:
@@ -107,9 +109,10 @@ class BaseGeometry(abc.ABC):
             self.parameters.update(parameters)
 
         self.mechanics_mesh = mechanics_mesh
-        self.ep_mesh = ep_mesh
-
         self.ffun = ffun
+        self.ep_mesh = ep_mesh
+        self.ffun_ep = ffun_ep
+
         self.microstructure = microstructure
         self.microstructure_ep = microstructure_ep
         self.validate()
@@ -149,6 +152,12 @@ class BaseGeometry(abc.ABC):
                 is_meshfunction=True,
                 dim=2,
                 mesh_key="mesh",
+            ),
+            "ffun_ep": H5Path(
+                h5group="/geometry/meshfunctions/ffun_ep",
+                is_meshfunction=True,
+                dim=2,
+                mesh_key="ep_mesh",
             ),
             "f0": H5Path(
                 h5group="/geometry/microstructure/mechanics/f0",
@@ -193,8 +202,8 @@ class BaseGeometry(abc.ABC):
     @abc.abstractmethod
     def _default_microstructure(
         self,
-        fiber_space: str,
         mesh: dolfin.Mesh,
+        ffun: dolfin.MeshFunction,
     ) -> pulse.Microstructure:
         ...
 
@@ -248,12 +257,10 @@ class BaseGeometry(abc.ABC):
     def _get_microstructure_if_None(
         self,
         mesh: dolfin.Mesh,
+        ffun: dolfin.MeshFunction,
         label: str,
     ) -> pulse.Microstructure:
-        microstructure = self._default_microstructure(
-            fiber_space=self.parameters["fiber_space"],
-            mesh=mesh,
-        )
+        microstructure = self._default_microstructure(mesh=mesh, ffun=ffun)
 
         if self.outdir is not None:
             path = self.outdir / f"microstructure{label}.h5"
@@ -273,17 +280,21 @@ class BaseGeometry(abc.ABC):
 
     @microstructure_ep.setter
     def microstructure_ep(self, microstructure: Optional[pulse.Microstructure]) -> None:
-        if microstructure is None:
+        if microstructure is None or microstructure.f0 is None:
             microstructure = self._interpolate_microstructure()
         self._microstructure_ep = microstructure
 
     def _interpolate_microstructure(self) -> pulse.Microstructure:
         element = self.f0.ufl_element()
-        V = dolfin.FunctionSpace(self.ep_mesh, element)
-        f0 = dolfin.interpolate(self.f0, V)
-        s0 = dolfin.interpolate(self.s0, V)
-        n0 = dolfin.interpolate(self.n0, V)
-        return pulse.Microstructure(f0=f0, s0=s0, n0=n0)
+        if element.family() == "Quadrature":
[…]
