### 2022-11-14 ComputationalPhysiology/simcardems#85 @jorgensd
@mbarzegary This has now been patched in conda: https://github.com/conda-forge/fenics-feedstock/pull/169#issuecomment-1313729111
Did you use conda when you got this issue, or did you install from source?
