+++ _sources/ventricular_geometry.md
@@ -1,6 +1,8 @@
-# Simple demo
+# Left ventricular geometry
 
-In this demo we show the most simple usage of the `simcardems` library using the python API
+In this demo we will run the same simple demo but using a ventricular geometry instead.
+Note that we currently only support a fixed base boundary condition, but plan to add support for
+more complicated boundary conditions in the future.
 
 Import the necessary libraries
 
@@ -17,18 +19,14 @@ import simcardems
 ```python
 # Create configurations with custom output directory
 here = Path(__file__).absolute().parent
-outdir = here / "results_simple_demo"
+outdir = here / "results_lv_ellipsoid"
 ```
-
-Specify paths to the geometry that we will use
 
 ```python
-geometry_path = here / "geometries/slab.h5"
-geometry_schema_path = here / "geometries/slab.json"
+# Specify paths to the geometry that we will use
+geometry_path = here / "geometries/lv_ellipsoid.h5"
+geometry_schema_path = here / "geometries/lv_ellipsoid.json"
 ```
-
-Please see https://computationalphysiology.github.io/cardiac_geometries/ for more info about the geometries
-
 
 ```python
 config = simcardems.Config(
@@ -90,28 +88,32 @@ This will create the output directory `results_simple_demo` with the following o
 ```
 results_simple_demo
 ├── results.h5
+├── results.json
 ├── state.h5
+├── state.json
 ```
 The file `state.h5` contains the final state which can be used if you want use the final state as a starting point for the next simulation.
-The file `results.h5` contains the Displacement ($u$), active tension ($T_a$), voltage ($V$) and calcium ($Ca$) for each time step.
-We can also plot the traces using the postprocess module
+The file `results.h5` contains different traces such as Displacement ($u$), active tension ($T_a$), voltage ($V$) and calcium ($Ca$) for each time step.
+The `.json` files are used to load load the geometries and, see https://computationalphysiology.github.io/cardiac_geometries/quickstart.html#schema for more info.
+We can also plot the traces using the postprocess module.
+
 
 
 
 ```python
-simcardems.postprocess.plot_state_traces(outdir.joinpath("results.h5"), "center")
+simcardems.postprocess.plot_state_traces(outdir.joinpath("results.h5"))
 ```
 
 
-Here we also specify that we want the trace from the center of the slab
+Here the traces generated by averaging over this domain.
 
 
 
-This will create a figure in the output directory called `state_traces.png` which in this case is shown in {numref}`Figure {number} <simple_demo_state_traces>` we see the resulting state traces, and can also see the instant drop in the active tension ($T_a$) at the time of the triggered release.
+This will create a figure in the output directory called `state_traces.png` which in this case is shown in {numref}`Figure {number} <lv_demo_state_traces>` we see the resulting state traces, and can also see the instant drop in the active tension ($T_a$) at the time of the triggered release.
 
-```{figure} figures/simple_demo_state_traces.png
+```{figure} figures/lv_demo_state_traces.png
 ---
-name: simple_demo_state_traces
+name: lv_demo_state_traces
 ---
 Traces of the stretch ($\lambda$), the active tension ($T_a$), the membrane potential ($V$) and the intercellular calcium concentration ($Ca$) at the center of the geometry.
 ```
@@ -126,11 +128,11 @@ simcardems.postprocess.make_xdmffiles(outdir.joinpath("results.h5"))
 ```
 
 
-The `xdmf` files are can be opened in [Paraview](https://www.paraview.org/download/) to visualize the different variables such as in {numref}`Figure {number} <simple-demo-paraview>`.
[…]
