+++ src/simcardems/version.py
@@ -1,1 +1,1 @@
-__version__ = "2022.3.2"
+__version__ = "2022.4.0"
