+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 2022.3.2
+current_version = 2022.4.0
 commit = True
 tag = True
 
