### 2022-11-15 1f93868
Attempt to fix tags in docker image
