+++ .github/workflows/docker-image.yml
@@ -52,5 +52,7 @@ jobs:
           context: .
           push: true
           platforms: linux/amd64,linux/arm64
-          tags: ${{ steps.meta.outputs.tags }} , latest
+          tags: |
+            ${{ steps.meta.outputs.tags }}
+            ghcr.io/computationalphysiology/cardiac_geometries:latest
           labels: ${{ steps.meta.outputs.labels }}
