### 2022-11-15 ComputationalPhysiology/simcardems#85 @mbarzegary
Great!
Yes, I installed FEniCS inside a conda environment.
