### 2022-11-15 ComputationalPhysiology/simcardems#85 @jorgensd
> Great! Yes, I installed FEniCS inside a conda environment.

Then it should hopefully work now with the latest conda builds: https://anaconda.org/conda-forge/fenics/files
