### 2022-11-15 76df75b
Attempt to fix tags in docker image - skip latest tag
