### 2022-11-17 0da3951
Correct boundary condition markers for non-default settings
