### 2022-11-17 3b1d61b
Add fix_right_plane to cli and use append instead of extend for BC
