+++ demos/drug_effects_demo.py
@@ -1,0 +1,114 @@
+# -*- coding: utf-8 -*-
+# # Drug effects demo
+#
+# In this demo we show how to modulate the cell model with pharmacological effects
+#
+#
+
+import pprint
+from pathlib import Path
+
+import simcardems
+
+# Create configurations with custom output directory
+here = Path(__file__).absolute().parent
+outdir = here / "results_drug_demo"
+
+# Specify paths to the geometry that we will use
+geometry_path = here / "geometries/slab.h5"
+geometry_schema_path = here / "geometries/slab.json"
+
+# Please see https://computationalphysiology.github.io/cardiac_geometries/ for more info about the geometries
+
+# Specify path to the file containing drug scaling factors
+drug_factors_path = here / "drug_factors/drug_factors.json"
+
+# The scaling factors in this file are applied to the corresponding ion currents as a modulator of the ion channel conductance. The example file reduces the sodium current to 10%.
+
+config = simcardems.Config(
+    outdir=outdir,
+    geometry_path=geometry_path,
+    geometry_schema_path=geometry_schema_path,
+    T=1000,
+    drug_factors_file=drug_factors_path,
+)
+
+
+# This will set :
+#
+# ```
+# {'PCL': 1000,
+#  'T': 1000,
+#  'bnd_rigid': False,
+#  'cell_init_file': '',
+#  'disease_state': 'healthy',
+#  'drug_factors_file': '',
+#  'dt': 0.05,
+#  'ep_ode_scheme': 'GRL1',
+#  'ep_preconditioner': 'sor',
+#  'ep_theta': 0.5,
+#  'fix_right_plane': False,
+#  'geometry_path': 'demos/geometries/slab.h5',
+#  'geometry_schema_path': 'demos/geometries/slab.json',
+#  'linear_mechanics_solver': 'mumps',
+#  'load_state': False,
+#  'loglevel': 20,
+#  'mechanics_ode_scheme': <Scheme.analytic: 'analytic'>,
+#  'mechanics_use_continuation': False,
+#  'mechanics_use_custom_newton_solver': False,
+#  'num_refinements': 1,
+#  'outdir': PosixPath('results_drug_demo'),
+#  'popu_factors_file': 'demos/drug_factors/drug_factors.json',
+#  'pre_stretch': None,
+#  'save_freq': 1,
+#  'set_material': '',
+#  'show_progress_bar': True,
+#  'spring': None,
+#  'traction': None}
+# ```
+
+
+# Print current configuration
+pprint.pprint(config.as_dict())
+
+runner = simcardems.Runner(config)
+runner.solve(T=config.T, save_freq=config.save_freq, show_progress_bar=True)
+
+
+# This will create the output directory `results_drug_demo` with the following output
+#
+# ```
+# results_simple_demo
+# ├── results.h5
+# ├── state.h5
+# ```
+# The file `state.h5` contains the final state which can be used if you want use the final state as a starting point for the next simulation.
+# The file `results.h5` contains the Displacement ($u$), active tension ($T_a$), voltage ($V$) and calcium ($Ca$) for each time step.
+# We can also plot the traces using the postprocess module
+#
+#
+
+simcardems.postprocess.plot_state_traces(outdir.joinpath("results.h5"), "center")
+
+#
+# Here we also specify that we want the trace from the center of the slab
+#
+
+# This will create a figure in the output directory called `state_traces.png` which in this case is shown in {numref}`Figure {number} <drug_demo_state_traces>` we see the resulting state traces.
+#
+# ```{figure} figures/drug_demo_state_traces.png
+# ---
+# name: drug_demo_state_traces
+# ---
[…]
