### 2022-11-21 483c6dd
Add demo for initial conditions
