### 2022-11-21 34c8762
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pre-commit/mirrors-mypy: v0.982 → v0.991](https://github.com/pre-commit/mirrors-mypy/compare/v0.982...v0.991)
- [github.com/streetsidesoftware/cspell-cli: v6.14.0 → v6.14.1](https://github.com/streetsidesoftware/cspell-cli/compare/v6.14.0...v6.14.1)
