### 2022-11-22 69d3d82
Remove example config output from demos
