+++ demos/initial_conditions_demo.py
@@ -31,40 +31,6 @@ config = simcardems.Config(
 )
 
 
-# This will set :
-#
-# ```
-# {'PCL': 1000,
-#  'T': 1000,
-#  'bnd_rigid': False,
-#  'cell_init_file': 'demos/initial_conditions/init_5000beats.json',
-#  'disease_state': 'healthy',
-#  'drug_factors_file': '',
-#  'dt': 0.05,
-#  'ep_ode_scheme': 'GRL1',
-#  'ep_preconditioner': 'sor',
-#  'ep_theta': 0.5,
-#  'fix_right_plane': False,
-#  'geometry_path': 'demos/geometries/slab.h5',
-#  'geometry_schema_path': 'demos/geometries/slab.json',
-#  'linear_mechanics_solver': 'mumps',
-#  'load_state': False,
-#  'loglevel': 20,
-#  'mechanics_ode_scheme': <Scheme.analytic: 'analytic'>,
-#  'mechanics_use_continuation': False,
-#  'mechanics_use_custom_newton_solver': False,
-#  'num_refinements': 1,
-#  'outdir': PosixPath('results_IC_demo'),
-#  'popu_factors_file': '',
-#  'pre_stretch': None,
-#  'save_freq': 1,
-#  'set_material': '',
-#  'show_progress_bar': True,
-#  'spring': None,
-#  'traction': None}
-# ```
-
-
 # Print current configuration
 pprint.pprint(config.as_dict())
 
