### 2022-11-22 27a9ad5
Change drug and IC input to paths instead of strings
