### 2022-11-23 ab057cf
Add same command for drug effect demo
