+++ src/simcardems/__init__.py
@@ -67,7 +67,7 @@ for module in ["matplotlib", "h5py", "FFC", "UFL"]:
 _dolfin.parameters["form_compiler"]["cpp_optimize"] = True
 flags = ["-O3", "-ffast-math"]
 _dolfin.parameters["form_compiler"]["cpp_optimize_flags"] = " ".join(flags)
-_dolfin.parameters["form_compiler"]["quadrature_degree"] = 3
+_dolfin.parameters["form_compiler"]["quadrature_degree"] = 4
 _dolfin.parameters["form_compiler"]["representation"] = "uflacs"
 
 _warnings.simplefilter("once", _QuadratureRepresentationDeprecationWarning)
