### 2022-11-24 ComputationalPhysiology/simcardems#89 @finsberg
We have now addressed all of these issues. Feel free to re-open this if needed.
