### 2022-11-24 ComputationalPhysiology/simcardems#84 @finsberg
Hermengild Arevalo and Samuel Wall are non-code contributors to the project. They have mainly been involved in scientific discussions and design choices of the software. Please let us know if this is not sufficient and if we should add this information to the paper.
