### 2022-11-24 5bdf200
Adding some more background on heart modeling and adding some more references to other softwares
