+++ joss/paper.bib
@@ -73,3 +73,42 @@
   year={2017},
   publisher={Elsevier}
 }
+
+@article{updegrove2017simvascular,
+  title={SimVascular: an open source pipeline for cardiovascular simulation},
+  author={Updegrove, Adam and Wilson, Nathan M and Merkow, Jameson and Lan, Hongzhi and Marsden, Alison L and Shadden, Shawn C},
+  journal={Annals of biomedical engineering},
+  volume={45},
+  number={3},
+  pages={525--541},
+  year={2017},
+  publisher={Springer}
+}
+
+@article{plank2021opencarp,
+  title={The openCARP simulation environment for cardiac electrophysiology},
+  author={Plank, Gernot and Loewe, Axel and Neic, Aurel and Augustin, Christoph and Huang, Yung-Lin and Gsell, Matthias AF and Karabelas, Elias and Nothstein, Mark and Prassl, Anton J and S{\'a}nchez, Jorge and others},
+  journal={Computer Methods and Programs in Biomedicine},
+  volume={208},
+  pages={106223},
+  year={2021},
+  publisher={Elsevier}
+}
+
+@article{bradley2011opencmiss,
+  title={OpenCMISS: a multi-physics \& multi-scale computational infrastructure for the VPH/Physiome project},
+  author={Bradley, Chris and Bowery, Andy and Britten, Randall and Budelmann, Vincent and Camara, Oscar and Christie, Richard and Cookson, Andrew and Frangi, Alejandro F and Gamage, Thiranja Babarenda and Heidlauf, Thomas and others},
+  journal={Progress in biophysics and molecular biology},
+  volume={107},
+  number={1},
+  pages={32--47},
+  year={2011},
+  publisher={Elsevier}
+}
+
+@article{cooper2020chaste,
+  title={Chaste: cancer, heart and soft tissue environment},
+  author={Cooper, Fergus R and Baker, Ruth E and Bernabeu, Miguel O and Bordas, Rafel and Bowler, Louise and Bueno-Orovio, Alfonso and Byrne, Helen M and Carapella, Valentina and Cardone-Noott, Louie and Cooper, Jonathan and others},
+  journal={Journal of Open Source Software},
+  year={2020}
+}
