+++ joss/paper.md
@@ -30,15 +30,15 @@ bibliography: paper.bib
 
 # Summary
 
-Modeling and simulations can help us to develop better understanding of the physiology and pathophysiology of the heart. Increased mechanicstic insight will enable us to improve and tailor treatment design strategies at the level of the population and individual.
+Cardiac modeling and simulations can help us to develop better understanding of the physiology and pathophysiology of the heart. Increased mechanicstic insight will enable us to improve and tailor treatment design strategies at the level of the population and individual.
 
-The heart consist of excitable tissue that responds to an electrical stimulus originating from the pacemaker cells of the heart. The excitation propagates through the whole heart and triggers contraction, which makes your heart contract and do its primary job to pump blood to the body. The electrical signal and contraction of the heart are tightly coupled and highly coordinated to maintain a steady heart beat. However, if something is wrong in the electrical signal, it might manifest itself in an irregular contraction. On the contrary, when the heart tissue stretches or contracts irregurlarly, there are mechanisms at the cellular level that will alter the electrical signal.
+The heart consist of excitable tissue that responds to an electrical stimulus originating from the pacemaker cells of the heart. The excitation propagates through the whole heart and triggers contraction, which makes the heart contract and do its primary job to pump blood to the body. The electrical signal and contraction of the heart are tightly coupled and highly coordinated to maintain a steady heart beat. However, if something is wrong in the electrical signal, it might manifest itself in an irregular contraction. On the contrary, when the heart tissue stretches or contracts irregurlarly, there are mechanisms at the cellular level that will alter the electrical signal.
 
-Modeling these effects in both directions requires a model that incorporates the interactions between the electrical and the mechanical components of the heart. Such a model is called an electro-mechanical model.
+Modeling these interactions in both directions requires a fully coupled model that incorporates the strong coupling between the electrical and the mechanical components of the heart. Such a model is called an electro-mechanical model.
 
 In addition to understanding cardiac mechanisms, an electro-mechanical model could be used to simulate the effect of drugs. Computational models of cardiac electrophysiology are already used to assess drug effects, but the interaction with mechanics is often ignored or simplified. A drug that blocks an ion channel in the cell, causes a change in the flow of ions across the cell membrane. This change in membrane voltage affects the propagation of the excitation wave through the heart. Through this change at the cell level, the drug indirectly affects the contraction of the heart. Understanding how the drug effect propagates to the organ level and affects the electrophysiology and mechanics of the heart is key to understand if a drug is safe and effective.
 
-An electromechanics solver contains a model that couples the different mechanistic models at the cellular level to the models at the tissue level, i.e a monodomain model for the electrophysiology and a hyperelastic continuum model for the mechanics.
+An electro-mechanical solver contains a model that couples the different models at the cellular level to the models at the tissue level, i.e a monodomain model for the electrophysiology and a hyperelastic continuum model for the mechanics.
 
 
 # Statement of need
@@ -51,9 +51,9 @@ For example SimVascular[@updegrove2017simvascular] is a software package that pr
 
 The OpenCarp[@plank2021opencarp] software focuses more on modeling the electrophysiology in the heart, while OpenCMISS[@bradley2011opencmiss] is more focused on cardiac mechanics.
 
[…]
