+++ joss/paper.md
@@ -30,13 +30,13 @@ bibliography: paper.bib
 
 # Summary
 
-Cardiac modeling and simulations can help us to develop better understanding of the physiology and pathophysiology of the heart. Increased mechanicstic insight will enable us to improve and tailor treatment design strategies at the level of the population and individual.
+Cardiac modeling and simulations are important tools to aid in understanding the physiology and pathophysiology of the heart. Improved mechanistic insight into the function and dysfunction of cardiac systems will enable scientists to develop novel methods and improve on general strategies for treatment of cardiac disease and injury, and to allow for clinical tailoring of treatment to the level of the individual.
 
-The heart consist of excitable tissue that responds to an electrical stimulus originating from the pacemaker cells of the heart. The excitation propagates through the whole heart and triggers contraction, which makes the heart contract and do its primary job to pump blood to the body. The electrical signal and contraction of the heart are tightly coupled and highly coordinated to maintain a steady heart beat. However, if something is wrong in the electrical signal, it might manifest itself in an irregular contraction. On the contrary, when the heart tissue stretches or contracts irregurlarly, there are mechanisms at the cellular level that will alter the electrical signal.
+The heart consist of electrically excitable tissue that responds to a stimulus originating from  pacemaker cells located in the atria. This excitation electrically propagates through the tissue of atria and then down into and through the ventricles, and drives synchronized muscle contraction that pumps blood throughout the body. The propagating electrical signal and and the resulting contraction of the heart tissues are tightly coupled and highly coordinated to allow the pump to work efficiently and react to changing needs of the body.  However, this tight coupling also can drive pathology, as for example if something is wrong in the electrical mechanisms, it might manifest as an irregular contraction. Or, when the heart tissue stretches or contracts irregularly, there are mechanisms at the cellular level that can alter the electrical behavior and drive dangerous arrhythmias.
 
-Modeling these interactions in both directions requires a fully coupled model that incorporates the strong coupling between the electrical and the mechanical components of the heart. Such a model is called an electro-mechanical model.
+Accounting for these intertwined interactions properly in a simulation of cardiac behavior requires a fully coupled model that incorporates the strong coupling between the active electrophysiological and the contractile mechanisms of the myocardial cells that the heart consists of, as well as the passive mechanical and electrical behavior of the overall cardiac tissue. Such a model is called an electro-mechanical model.
 
-In addition to understanding cardiac mechanisms, an electro-mechanical model could be used to simulate the effect of drugs. Computational models of cardiac electrophysiology are already used to assess drug effects, but the interaction with mechanics is often ignored or simplified. A drug that blocks an ion channel in the cell, causes a change in the flow of ions across the cell membrane. This change in membrane voltage affects the propagation of the excitation wave through the heart. Through this change at the cell level, the drug indirectly affects the contraction of the heart. Understanding how the drug effect propagates to the organ level and affects the electrophysiology and mechanics of the heart is key to understand if a drug is safe and effective.
+One prominent but unexplored use of electro-mechanical modelling is the simulation of the effect of cardioactive drugs. Computational models of cardiac electrophysiology are already heavily used to assess drug effects of the electrical dynamics of the heart, but the interaction with overall mechanics is often ignored or simplified. As drug induced changes to cardiac electrophysiology, such as a block of a key ion channel, can propagate through the tightly coupled connection and indirectly affect the contraction of the heart, understanding this process and how compounds may alter both the electrophysiology and mechanics of the heart is key to understand if a drug is safe and effective.
 
 An electro-mechanical solver contains a model that couples the different models at the cellular level to the models at the tissue level, i.e a monodomain model for the electrophysiology and a hyperelastic continuum model for the mechanics.
 
