+++ joss/paper.md
@@ -30,20 +30,22 @@ bibliography: paper.bib
 
 # Summary
 
-Cardiac modeling and simulations are important tools to aid in understanding the physiology and pathophysiology of the heart. Improved mechanistic insight into the function and dysfunction of cardiac systems will enable scientists to develop novel methods and improve on general strategies for treatment of cardiac disease and injury, and to allow for clinical tailoring of treatment to the level of the individual.
+The heart is composed of electrically excitable tissues that contract upon stimulation.  During a normal beat, the inital stimulus to drive the heart originates from a group of pacemaker cells located in the atria, and electrically propagates first through the atria and then down into and through the ventricles.  This travelling stimulus activates the myocytes as it arrives, driving a cascade of cellular processes that end in the release of calcium from intracellular stores.  This caclium then causes the activation and cycling of actin - myosin cross bridges within the mycotyes, creating the muscle contraction that pumps blood throughout the body. The propagating electrical signal and and the resulting contraction of the heart tissues are tightly coupled and highly coordinated to allow the pump to work efficiently and react to changing needs of the body.  However, this tight coupling also can drive pathology.  For example if something is wrong in the electrical processes, it might manifest as an irregular contraction. Or, when the heart tissue is stretched or contracts irregularly, there are mechanisms at the cellular level that can alter the electrical behavior and drive dangerous arrhythmias.
 
-The heart consist of electrically excitable tissue that responds to a stimulus originating from  pacemaker cells located in the atria. This excitation electrically propagates through the tissue of atria and then down into and through the ventricles, and drives synchronized muscle contraction that pumps blood throughout the body. The propagating electrical signal and and the resulting contraction of the heart tissues are tightly coupled and highly coordinated to allow the pump to work efficiently and react to changing needs of the body.  However, this tight coupling also can drive pathology, as for example if something is wrong in the electrical mechanisms, it might manifest as an irregular contraction. Or, when the heart tissue stretches or contracts irregularly, there are mechanisms at the cellular level that can alter the electrical behavior and drive dangerous arrhythmias.
+Modeling and simulation are used extensively to describe the physics of these electrical and mehcanical processes and how they are connected.   Accounting for the intertwined interactions properly in a simulation of cardiac behavior requires a complex model that incorporates the strong coupling between the active electrophysiological and the contractile mechanisms of the myocardial cells, as well as the passive mechanical and electrical behavior of the overall cardiac tissue. Such a model is referred to as an electro-mechanical model.  Typically, an electro-mechanical solver contains a number of sub models that describe and connect the different physics across the relevant scales, e.g. a monodomain model for the electrophysiology coupled to and a hyperelastic continuum model for the mechanics.
 
-Accounting for these intertwined interactions properly in a simulation of cardiac behavior requires a fully coupled model that incorporates the strong coupling between the active electrophysiological and the contractile mechanisms of the myocardial cells that the heart consists of, as well as the passive mechanical and electrical behavior of the overall cardiac tissue. Such a model is called an electro-mechanical model.
[…]
