### 2022-12-01 37fb70b
More fixes to paper
