+++ joss/paper.bib
@@ -112,3 +112,14 @@
   journal={Journal of Open Source Software},
   year={2020}
 }
+
+@article{sundnes2014improved,
+  title     = {Improved discretisation and linearisation of active tension in strongly coupled cardiac electro-mechanics simulations},
+  author    = {Sundnes, Joakim and Wall, S and Osnes, Harald and Thorvaldsen, Tom and McCulloch, Andrew D},
+  journal   = {Computer methods in biomechanics and biomedical engineering},
+  volume    = {17},
+  number    = {6},
+  pages     = {604--615},
+  year      = {2014},
+  publisher = {Taylor \& Francis}
+}
