### 2022-12-02 ComputationalPhysiology/simcardems#88 @finsberg
@mbarzegary We have now update the paper as well as provided a demo using a more complex geometry. I am closing this now, but feel free to re-open if you are not happy with the changes we made.
