+++ joss/paper.bib
@@ -75,6 +75,7 @@
 }
 
 @article{updegrove2017simvascular,
+  doi={10.1007/s10439-016-1762-8},
   title={SimVascular: an open source pipeline for cardiovascular simulation},
   author={Updegrove, Adam and Wilson, Nathan M and Merkow, Jameson and Lan, Hongzhi and Marsden, Alison L and Shadden, Shawn C},
   journal={Annals of biomedical engineering},
@@ -107,6 +108,7 @@
 }
 
 @article{cooper2020chaste,
+  doi={10.21105/joss.01848},
   title={Chaste: cancer, heart and soft tissue environment},
   author={Cooper, Fergus R and Baker, Ruth E and Bernabeu, Miguel O and Bordas, Rafel and Bowler, Louise and Bueno-Orovio, Alfonso and Byrne, Helen M and Carapella, Valentina and Cardone-Noott, Louie and Cooper, Jonathan and others},
   journal={Journal of Open Source Software},
@@ -114,6 +116,7 @@
 }
 
 @article{sundnes2014improved,
+  doi={10.1080/10255842.2012.704368},
   title     = {Improved discretisation and linearisation of active tension in strongly coupled cardiac electro-mechanics simulations},
   author    = {Sundnes, Joakim and Wall, S and Osnes, Harald and Thorvaldsen, Tom and McCulloch, Andrew D},
   journal   = {Computer methods in biomechanics and biomedical engineering},
