### 2022-12-05 c050da4
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/streetsidesoftware/cspell-cli: v6.15.0 → v6.16.0](https://github.com/streetsidesoftware/cspell-cli/compare/v6.15.0...v6.16.0)
