### 2022-12-06 75ff4e3
Add a lot more debug statements and save residuals from mechancis model in debug mode
