### 2022-12-06 7f96b17
Make sure debug-mode is passed correcly from the cli
