+++ src/simcardems/cli.py
@@ -177,6 +177,7 @@ def run(
     show_progress_bar: bool,
     save_freq: int,
     loglevel: int,
+    debug_mode: bool,
     num_refinements: int,
     set_material: str,
     drug_factors_file: utils.PathLike,
@@ -205,6 +206,7 @@ def run(
         show_progress_bar=show_progress_bar,
         save_freq=save_freq,
         loglevel=loglevel,
+        debug_mode=debug_mode,
         num_refinements=num_refinements,
         set_material=set_material,
         drug_factors_file=drug_factors_file,
