### 2022-12-06 4c5bc92
Change type of residual for datacollector
