### 2022-12-08 e61c9ae
Apply suggestions from code review
Co-authored-by: Henrik Finsberg <henriknf@simula.no>
