### 2022-12-08 c4c736a
Apply suggestions from code review
Co-authored-by: Henrik Finsberg <henriknf@simula.no>
