### 2022-12-08 e36f755
Apply suggestions from code review
