### 2022-12-08 8207abf
Apply suggestions from code review
