### 2022-12-08 866233a
Apply suggestions from code review
