### 2022-12-08 ee0b34a
Apply suggestions from code review
