### 2022-12-08 bf4c3a5
Misc fixes
