+++ setup.cfg
@@ -74,7 +74,7 @@ select = B,C,E,F,W,T4
 test = pytest
 
 [tool:pytest]
-addopts = --cov=src/simcardems --cov-report html --cov-report xml --cov-report term-missing -v
+addopts = --cov=simcardems --cov-report html --cov-report xml --cov-report term-missing -v
 testpaths =
     tests
 markers =
