### 2022-12-08 81e0eae
Incorporate suggestion from jørgen
