+++ .cspell_dict.txt
@@ -264,6 +264,7 @@ minlmbda
 mL
 mode
 modelparams
+modifyitems
 monodomain
 mpio
 multigrid
