+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 2022.5.0
+current_version = 2022.5.1
 commit = True
 tag = True
 
