### 2022-12-12 5581704
Begin work to make the coupling explicit
