+++ src/simcardems/setup_models.py
@@ -12,12 +12,13 @@ from . import config
 from . import em_model
 from . import ep_model
 from . import geometry
-from . import land_model
 from . import mechanics_model
 from . import save_load_functions as io
 from . import utils
 from .em_model import EMCoupling
 from .ORdmm_Land import ORdmm_Land as CellModel
+
+# from . import land_model
 
 logger = utils.getLogger(__name__)
 
@@ -62,7 +63,7 @@ def setup_EM_model(config: config.Config):
         spring=config.spring,
         fix_right_plane=config.fix_right_plane,
         set_material=config.set_material,
-        mechanics_ode_scheme=config.mechanics_ode_scheme,
+        # mechanics_ode_scheme=config.mechanics_ode_scheme,
         use_custom_newton_solver=config.mechanics_use_custom_newton_solver,
         debug_mode=config.debug_mode,
     )
@@ -86,13 +87,9 @@ def setup_mechanics_solver(
     spring: typing.Union[dolfin.Constant, float] = None,
     fix_right_plane: bool = config.Config.fix_right_plane,
     debug_mode: bool = config.Config.debug_mode,
-    mechanics_ode_scheme: land_model.Scheme = config.Config.mechanics_ode_scheme,
     set_material: str = "",
     linear_solver="mumps",
     use_custom_newton_solver: bool = config.Config.mechanics_use_custom_newton_solver,
-    Zetas_prev=None,
-    Zetaw_prev=None,
-    lmbda_prev=None,
     state_prev=None,
 ):
     """Setup mechanics model with dirichlet boundary conditions or rigid motion."""
@@ -110,25 +107,33 @@ def setup_mechanics_solver(
         b_fs=0.0,
     )
 
-    active_model = land_model.LandModel(
+    V = dolfin.FunctionSpace(geo.mesh, "CG", 1)
+    Ta = dolfin.Function(V)
+    # active_model = land_model.LandModel(
+    #     f0=geo.f0,
+    #     s0=geo.s0,
+    #     n0=geo.n0,
+    #     eta=0,
+    #     parameters=cell_params,
+    #     XS=coupling.XS_mech,
+    #     XW=coupling.XW_mech,
+    #     mesh=coupling.mech_mesh,
+    #     scheme=mechanics_ode_scheme,
+    #     Zetas=Zetas_prev,
+    #     Zetaw=Zetaw_prev,
+    #     lmbda=lmbda_prev,
+    # )
+
+    material_kwargs = dict(
+        active_model="active_stress",
+        activation=Ta,
         f0=geo.f0,
         s0=geo.s0,
         n0=geo.n0,
-        eta=0,
-        parameters=cell_params,
-        XS=coupling.XS_mech,
-        XW=coupling.XW_mech,
-        mesh=coupling.mech_mesh,
-        scheme=mechanics_ode_scheme,
-        Zetas=Zetas_prev,
-        Zetaw=Zetaw_prev,
-        lmbda=lmbda_prev,
-    )
-
-    material = pulse.HolzapfelOgden(
-        active_model=active_model,
         parameters=material_parameters,
     )
+
+    material = pulse.HolzapfelOgden(**material_kwargs)
 
     if set_material == "Guccione":
         material_parameters = pulse.Guccione.default_parameters()
@@ -137,10 +142,7 @@ def setup_mechanics_solver(
         material_parameters["bfs"] = 4.0
         material_parameters["bt"] = 2.0
 
-        material = pulse.Guccione(
-            params=material_parameters,
-            active_model=active_model,
-        )
+        material = pulse.Guccione(**material_kwargs)
 
     problem = mechanics_model.create_problem(
         material=material,
[…]
