+++ demos/simple_demo.py
@@ -27,7 +27,7 @@ config = simcardems.Config(
     outdir=outdir,
     geometry_path=geometry_path,
     geometry_schema_path=geometry_schema_path,
-    T=1000,
+    T=10,
 )
 
 
