+++ src/simcardems/postprocess.py
@@ -44,15 +44,28 @@ def extract_traces(loader: DataLoader, reduction: str = "average"):
 
     all_names = {
         "mechanics": [
-            "lmbda",
-            "Ta",
-            "Zetas_mech",
-            "Zetaw_mech",
-            "XS_mech",
-            "XW_mech",
+            # "lmbda",
+            # "Ta",
+            # "Zetas_mech",
+            # "Zetaw_mech",
+            # "XS_mech",
+            # "XW_mech",
             "u",
         ],
-        "ep": ["V", "Ca", "XS", "XW", "CaTrpn", "TmB", "Cd", "Zetas", "Zetaw"],
+        "ep": [
+            "V",
+            "Ca",
+            "XS",
+            "XW",
+            "CaTrpn",
+            "TmB",
+            "Cd",
+            "Zetas",
+            "Zetaw",
+            "lmbda",
+            "dLambda",
+            "Ta",
+        ],
     }
 
     values = {
@@ -81,13 +94,12 @@ def extract_traces(loader: DataLoader, reduction: str = "average"):
                     reduction=reduction,
                 )
 
-    values["mechanics"]["inv_lmbda"] = 1 - values["mechanics"]["lmbda"]
+    # values["mechanics"]["inv_lmbda"] = 1 - values["mechanics"]["lmbda"]
     return values
 
 
 def plot_state_traces(results_file: utils.PathLike, reduction: str = "average"):
-    fig, ax = plt.subplots(2, 2, figsize=(10, 8), sharex=True)
-    fig2, ax2 = plt.subplots(2, 4, figsize=(10, 8), sharex=True)
+
     results_file = Path(results_file)
     if not results_file.is_file():
         raise FileNotFoundError(f"File {results_file} does not exist")
@@ -105,77 +117,44 @@ def plot_state_traces(results_file: utils.PathLike, reduction: str = "average"):
             values["ep"]["Ca"],
             0.0002,
         )
+    fig, axs = plt.subplots(2, 2, figsize=(10, 8), sharex=True)
 
-    ax[0, 0].plot(times, values["mechanics"]["lmbda"])
-    ax[0, 1].plot(times, values["mechanics"]["Ta"])
-    ax[1, 0].plot(times, values["ep"]["V"])
-    ax[1, 1].plot(times, values["ep"]["Ca"])
+    for i, (group, key) in enumerate(
+        (("ep", "lmbda"), ("ep", "Ta"), ("ep", "V"), ("ep", "Ca")),
+    ):
+        ax = axs.flatten()[i]
+        y = values[group][key]
+        ax.plot(times, y)
+        if key == "lmbda":
+            ax.set_title(r"$\lambda$")
+            ax.set_ylim(min(0.9, min(y)), max(1.1, max(y)))
+        else:
+            ax.set_title(key)
+        ax.grid()
 
-    ax[0, 0].set_title(r"$\lambda$")
-    ax[0, 1].set_title("Ta")
-    ax[1, 0].set_title("V")
-    ax[1, 1].set_title("Ca")
-    for axi in ax.flatten():
-        axi.grid()
-        if False:
-            axi.set_xlim([0, 5000])
-    ax[1, 0].set_xlabel("Time [ms]")
-    ax[1, 1].set_xlabel("Time [ms]")
-    ax[0, 0].set_ylim(
-        [
-            min(0.9, min(values["mechanics"]["lmbda"][1:])),
-            max(1.1, max(values["mechanics"]["lmbda"][1:])),
-        ],
-    )
+    axs[1, 0].set_xlabel("Time [ms]")
[…]

[…]
