+++ src/simcardems/em_model.py
@@ -1,6 +1,9 @@
+from typing import Optional
+
 import dolfin
 
 from . import geometry as _geometry
+from . import land_model
 from . import utils
 from .mechanics_model import MechanicsProblem
 
@@ -12,30 +15,35 @@ class EMCoupling:
     def __init__(
         self,
         geometry: _geometry.BaseGeometry,
+        lmbda: Optional[dolfin.Function] = None,
     ) -> None:
         logger.debug("Create EM coupling")
         self.geometry = geometry
-        # self.XS_mech = dolfin.Function(self.V_mech, name="XS_mech")
-        # self.XW_mech = dolfin.Function(self.V_mech, name="XW_mech")
-        # self.lmbda_mech = dolfin.Function(self.V_mech, name="lambda_mech")
-        # self.Zetas_mech = dolfin.Function(self.V_mech, name="Zetas_mech")
-        # self.Zetaw_mech = dolfin.Function(self.V_mech, name="Zetaw_mech")
+
+        self.V_mech = dolfin.FunctionSpace(self.mech_mesh, "CG", 1)
+        self.XS_mech = dolfin.Function(self.V_mech, name="XS_mech")
+        self.XW_mech = dolfin.Function(self.V_mech, name="XW_mech")
+        self.Zetas_mech = dolfin.Function(self.V_mech, name="Zetas_mech")
+        self.Zetaw_mech = dolfin.Function(self.V_mech, name="Zetaw_mech")
 
         self.V_ep = dolfin.FunctionSpace(self.ep_mesh, "CG", 1)
-        self.W_ep = dolfin.VectorFunctionSpace(self.ep_mesh, "CG", 2)
-        # self.XS_ep = dolfin.Function(self.V_ep, name="XS_ep")
-        # self.XW_ep = dolfin.Function(self.V_ep, name="XW_ep")
+        self.XS_ep = dolfin.Function(self.V_ep, name="XS_ep")
+        self.XW_ep = dolfin.Function(self.V_ep, name="XW_ep")
         self.lmbda_ep = dolfin.Function(self.V_ep, name="lambda_ep")
-        self.lmbda_ep.vector()[:] = 1.0
         self.lmbda_ep_prev = dolfin.Function(self.V_ep, name="lambda_ep_prev")
-        self.lmbda_ep_prev.vector()[:] = 1.0
         self._dLambda = dolfin.Function(self.V_ep, name="dLambda")
-        self.Ta_ep = dolfin.Function(self.V_ep, name="Ta_ep")
-
         self.Zetas_ep = dolfin.Function(self.V_ep, name="Zetas_ep")
         self.Zetaw_ep = dolfin.Function(self.V_ep, name="Zetaw_ep")
 
+        self.W_ep = dolfin.VectorFunctionSpace(self.ep_mesh, "CG", 2)
         self.u_ep = dolfin.Function(self.W_ep, name="u_ep")
+
+        self._projector_V_ep = utils.Projector(self.V_ep)
+        self._projector_V_mech = utils.Projector(self.V_mech)
+
+        if lmbda is not None:
+            self.lmbda_ep.vector()[:] = lmbda.vector()
+            self.lmbda_ep_prev.vector()[:] = lmbda.vector()
 
     @property
     def mech_mesh(self):
@@ -47,32 +55,49 @@ class EMCoupling:
 
     @property
     def mech_state(self):
-        return self._mech_solver.state
+        return self.mech_solver.state
 
     @property
     def dLambda_ep(self):
         self._dLambda.vector()[:] = self.lmbda_ep.vector() - self.lmbda_ep_prev.vector()
         return self._dLambda
 
-    def _project_lmbda(self):
+    @property
[…]
