+++ demos/simple_demo.py
@@ -27,7 +27,8 @@ config = simcardems.Config(
     outdir=outdir,
     geometry_path=geometry_path,
     geometry_schema_path=geometry_schema_path,
-    T=10,
+    T=15,
+    load_state=True,
 )
 
 

[…]
