+++ demos/simple_demo.py
@@ -69,8 +69,8 @@ config = simcardems.Config(
 # Print current configuration
 pprint.pprint(config.as_dict())
 
-# runner = simcardems.Runner(config)
-# runner.solve(T=config.T, save_freq=config.save_freq, show_progress_bar=True)
+runner = simcardems.Runner(config)
+runner.solve(T=config.T, save_freq=config.save_freq, show_progress_bar=True)
 
 
 # This will create the output directory `results_simple_demo` with the following output
