### 2022-12-14 228ece3
Remove lambda and dlambda from states
