+++ src/simcardems/ORdmm_Land.py
@@ -2,8 +2,6 @@
 
 The module was autogenerated from a gotran ode file
 """
-from __future__ import division
-
 from collections import OrderedDict
 from typing import Dict
 
@@ -34,7 +32,13 @@ def vs_functions_to_dict(vs):
 
 
 class ORdmm_Land(CardiacCellModel):
-    def __init__(self, params=None, init_conditions=None):
+    def __init__(
+        self,
+        lmbda: dolfin.Function,
+        dLambda: dolfin.Function,
+        params=None,
+        init_conditions=None,
+    ):
         """
         Create cardiac cell model
 
@@ -46,6 +50,8 @@ class ORdmm_Land(CardiacCellModel):
         """
         logger.debug("Initialize ORdmm Land model")
         super().__init__(params, init_conditions)
+        self.lmbda = lmbda
+        self.dLambda = dLambda
 
     @staticmethod
     def default_parameters(disease_state: str = "healthy") -> Dict[str, float]:
@@ -304,8 +310,6 @@ class ORdmm_Land(CardiacCellModel):
                 ("cai", 0.0001),
                 ("Zetas", 0),
                 ("Zetaw", 0),
-                ("lmbda", 1),
-                ("dLambda", 0),
             ],
         )
         return ic
@@ -317,7 +321,7 @@ class ORdmm_Land(CardiacCellModel):
         time = time if time else Constant(0.0)
         logger.debug("Evaluate transmembrane current")
         # Assign states
-        assert len(s) == 49
+        assert len(s) == 47
         (
             CaMKt,
             m,
@@ -366,8 +370,6 @@ class ORdmm_Land(CardiacCellModel):
             cai,
             Zetas,
             Zetaw,
-            lmbda,
-            dLambda,
         ) = s
 
         # Assign parameters
@@ -771,15 +773,15 @@ class ORdmm_Land(CardiacCellModel):
         Gsac_k = (0.2882 * 800.0 / 210.0) * scale_drug_Isack  # Pueyo endo
 
         Isac_P_ns = ufl.conditional(
-            ufl.lt(lmbda, 1.0),
+            ufl.lt(self.lmbda, 1.0),
             0.0,
-            Gsac_ns * ((lmbda - 1.0) / (lambda_max - 1.0)) * (v - Esac_ns),
+            Gsac_ns * ((self.lmbda - 1.0) / (lambda_max - 1.0)) * (v - Esac_ns),
         )
         Isac_P_k = ufl.conditional(
-            ufl.lt(lmbda, 1.0),
+            ufl.lt(self.lmbda, 1.0),
             0.0,
             Gsac_k
-            * ((lmbda - 1.0) / (lambda_max - 1.0))
+            * ((self.lmbda - 1.0) / (lambda_max - 1.0))
             * (1.0 / (1.0 + ufl.exp((19.05 - v) / (29.98)))),
         )
 
@@ -829,7 +831,7 @@ class ORdmm_Land(CardiacCellModel):
         time = time if time else Constant(0.0)
 
         # Assign states
-        assert len(s) == 49
+        assert len(s) == 47
         (
             CaMKt,
             m,
@@ -878,8 +880,6 @@ class ORdmm_Land(CardiacCellModel):
             cai,
             Zetas,
             Zetaw,
-            lmbda,
-            dLambda,
         ) = s
 
         # Assign parameters
@@ -1040,7 +1040,7 @@ class ORdmm_Land(CardiacCellModel):
         HF_scaling_cat50_ref = self._parameters["HF_scaling_cat50_ref"]
 
         # Init return args
-        F_expressions = [dolfin.Constant(0.0)] * 49
[…]
