### 2022-12-14 5467666
Remove lmbda, Zetas and Zetaw from states
