+++ src/simcardems/ORdmm_Land.py
@@ -34,7 +34,14 @@ def vs_functions_to_dict(vs):
 
 
 class ORdmm_Land(CardiacCellModel):
-    def __init__(self, params=None, init_conditions=None):
+    def __init__(
+        self,
+        lmbda: dolfin.Function,
+        Zetas: dolfin.Function,
+        Zetaw: dolfin.Function,
+        params=None,
+        init_conditions=None,
+    ):
         """
         Create cardiac cell model
 
@@ -46,6 +53,9 @@ class ORdmm_Land(CardiacCellModel):
         """
         logger.debug("Initialize ORdmm Land model")
         super().__init__(params, init_conditions)
+        self.lmbda = lmbda
+        self.Zetas = Zetas
+        self.Zetaw = Zetaw
 
     @staticmethod
     def default_parameters(disease_state: str = "healthy") -> Dict[str, float]:
@@ -302,9 +312,6 @@ class ORdmm_Land(CardiacCellModel):
                 ("TmB", 1),
                 ("Cd", 0),
                 ("cai", 0.0001),
-                ("lmbda", 1),
-                ("Zetas", 0),
-                ("Zetaw", 0),
             ],
         )
         return ic
@@ -316,7 +323,7 @@ class ORdmm_Land(CardiacCellModel):
         time = time if time else Constant(0.0)
         logger.debug("Evaluate transmembrane current")
         # Assign states
-        assert len(s) == 48
+        assert len(s) == 45
         (
             CaMKt,
             m,
@@ -363,9 +370,6 @@ class ORdmm_Land(CardiacCellModel):
             TmB,
             Cd,
             cai,
-            lmbda,
-            Zetas,
-            Zetaw,
         ) = s
 
         # Assign parameters
@@ -769,15 +773,15 @@ class ORdmm_Land(CardiacCellModel):
         Gsac_k = (0.2882 * 800.0 / 210.0) * scale_drug_Isack  # Pueyo endo
 
         Isac_P_ns = ufl.conditional(
-            ufl.lt(lmbda, 1.0),
+            ufl.lt(self.lmbda, 1.0),
             0.0,
-            Gsac_ns * ((lmbda - 1.0) / (lambda_max - 1.0)) * (v - Esac_ns),
+            Gsac_ns * ((self.lmbda - 1.0) / (lambda_max - 1.0)) * (v - Esac_ns),
         )
         Isac_P_k = ufl.conditional(
-            ufl.lt(lmbda, 1.0),
+            ufl.lt(self.lmbda, 1.0),
             0.0,
             Gsac_k
-            * ((lmbda - 1.0) / (lambda_max - 1.0))
+            * ((self.lmbda - 1.0) / (lambda_max - 1.0))
             * (1.0 / (1.0 + ufl.exp((19.05 - v) / (29.98)))),
         )
 
@@ -827,7 +831,7 @@ class ORdmm_Land(CardiacCellModel):
         time = time if time else Constant(0.0)
 
         # Assign states
-        assert len(s) == 48
+        assert len(s) == 45
         (
             CaMKt,
             m,
@@ -874,9 +878,6 @@ class ORdmm_Land(CardiacCellModel):
             TmB,
             Cd,
             cai,
-            lmbda,
-            Zetas,
-            Zetaw,
         ) = s
 
         # Assign parameters
@@ -1040,7 +1041,7 @@ class ORdmm_Land(CardiacCellModel):
[…]
