### 2022-12-21 #4753 @sdelandtsheer
Hello,
I have not resumed the reviewing as I had trouble installing the tool on my machine. I had understood from the message on Oct16th someone else had plugged in for that part.
"With regards to one review not being able to fully test the submission, I'll step in and finalise some tests myself and review the software which, depending on the other review should be fine. I've discussed this with other editors and that's all okay-ed."
Is there something I should finalize? I do not have time for a full test until a few weeks.
