### 2022-12-28 #4753 @AoifeHughes
Just been over all the comments here from @mbarzegary, and I am happy having done a confirmatory review myself, I can recommend this for publication.

@finsberg the next steps are for you to:
1. Make a tagged release of the software
2. Deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/).
3. Comment the URL for that data-archive here.
4. Tag me with this info and the version number to be associated (if it has changed). I'll then proceed.

Thanks again Reviewers :)
