### 2023-01-02 9a99d61
Change where zeta and lambda get updated
