### 2023-01-02 3e898ff
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/streetsidesoftware/cspell-cli: v6.17.0 → v6.17.1](https://github.com/streetsidesoftware/cspell-cli/compare/v6.17.0...v6.17.1)
