+++ src/simcardems/version.py
@@ -1,1 +1,1 @@
-__version__ = "2022.5.1"
+__version__ = "2023.0.0"
