+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 2022.5.1
+current_version = 2023.0.0
 commit = True
 tag = True
 
