### 2023-01-04 #4753 @finsberg
@AoifeHughes I have now created a release for [version 2023.0.0](https://github.com/ComputationalPhysiology/simcardems/releases/tag/v2023.0.0) with a [specific tag (v2023.0.0)](https://github.com/ComputationalPhysiology/simcardems/tree/v2023.0.0) and also archived this at [Zenodo](https://zenodo.org/record/7503468). Badge for Zenodo is also added to the README in the repository.
