### 2023-01-09 #4753 @AoifeHughes
One minor thing I noticed is some references just don't have DOI's. I think they're book chapters? Is this something that can be added or is it in an acceptable state (@danielskatz 2nd op on this would be welcome, also, please note the issue with me taking over as a reviewer#2 hence the semi completed checklist for one reviewer.)

Other than that, thanks @finsberg, ab EiC will have a check now.
