+++ joss/paper.bib
@@ -1,4 +1,5 @@
 @book{logg2012automated,
+  doi       = {10.1007/978-3-642-23099-8},
   title     = {Automated solution of differential equations by the finite element method: The {FEniCS} book},
   author    = {Logg, Anders and Mardal, Kent-Andre and Wells, Garth},
   volume    = {84},
@@ -53,6 +54,7 @@
 }
 
 @article{o2011simulation,
+  doi={10.1371/journal.pcbi.1002061},
   title={Simulation of the undiseased human cardiac ventricular action potential: model formulation and experimental validation},
   author={O'Hara, Thomas and Vir{\'a}g, L{\'a}szl{\'o} and Varr{\'o}, Andr{\'a}s and Rudy, Yoram},
   journal={PLoS computational biology},
@@ -87,6 +89,7 @@
 }
 
 @article{plank2021opencarp,
+  doi={10.1016/j.cmpb.2021.106223},
   title={The openCARP simulation environment for cardiac electrophysiology},
   author={Plank, Gernot and Loewe, Axel and Neic, Aurel and Augustin, Christoph and Huang, Yung-Lin and Gsell, Matthias AF and Karabelas, Elias and Nothstein, Mark and Prassl, Anton J and S{\'a}nchez, Jorge and others},
   journal={Computer Methods and Programs in Biomedicine},
@@ -97,6 +100,7 @@
 }
 
 @article{bradley2011opencmiss,
+  doi={10.1016/j.pbiomolbio.2011.06.015},
   title={OpenCMISS: a multi-physics \& multi-scale computational infrastructure for the VPH/Physiome project},
   author={Bradley, Chris and Bowery, Andy and Britten, Randall and Budelmann, Vincent and Camara, Oscar and Christie, Richard and Cookson, Andrew and Frangi, Alejandro F and Gamage, Thiranja Babarenda and Heidlauf, Thomas and others},
   journal={Progress in biophysics and molecular biology},
