### 2023-01-09 #4753 @danielskatz
@AoifeHughes & @finsberg - JOSS practice is to provide DOIs for anything that has a DOI, but not for things that don't 🙂
