+++ joss/paper.md
@@ -36,7 +36,7 @@ Modeling and simulation are used extensively to describe the physics of these el
 
 Here we present a fully coupled electromechanical model in the context of evaluating the effects of cardioactive drugs on the heart. Our framework incorporates a modern human cell ordinary differential model (ODE) that describes the electrophysiology of the cell and how stimulation leads to internal calcium release, and how this calcium is coupled to cross-bridge dynamics and active force in the cell.  This 0D model is embedded in a monodomain partial differential equation (PDE) representation of the electrical substrate, as well as a corresponding incompressible, anisotropic, hyperelastic continuum model where contraction is governed by active stress.  The expansion and contraction of the continuum are governed by boundary conditions and the cellular calculations of calcium and cross-bridge states.  At the same time, change in the mechanical model is feedback to the cellular system, providing length and shortening velocity feedback to the ODE model.
 
-This strongly-coupled system is implemented in the open source finite element framework FEniCS[@logg2012automated] using an iterative conjugate gradient method with a geometric algebraic multigrid preconditioner to solve the monodomain equation, and a parallel sparse direct solver to solve the elasticity equations. The equations in the 0D ODE model are solved partly during the solving of the monodomain model and partly during solving of the mechanics to ensure the stability of the strongly coupled system[@sundnes2014improved].
+This strongly-coupled system is implemented in the open source finite element framework FEniCS [@logg2012automated] using an iterative conjugate gradient method with a geometric algebraic multigrid preconditioner to solve the monodomain equation, and a parallel sparse direct solver to solve the elasticity equations. The equations in the 0D ODE model are solved partly during the solving of the monodomain model and partly during solving of the mechanics to ensure the stability of the strongly coupled system [@sundnes2014improved].
 
 Several demos have been created, including demos using a slab-type geometry as well as one using a more realistic ellipsoidal geometry.
 
@@ -51,11 +51,11 @@ There have been significant developments and improvements in computational model
 
 There exist a variety of models used for studying the heart at different scales and several software packages provide different implementations of different models. Most software packages focus on one specific aspect of cardiac modeling.
 
-For example, SimVascular[@updegrove2017simvascular] is a software package that provides a complete pipeline from segmenting medical images of patients and using these generated geometries in blood flow simulations.
+For example, SimVascular [@updegrove2017simvascular] is a software package that provides a complete pipeline from segmenting medical images of patients and using these generated geometries in blood flow simulations.
 
-The OpenCarp[@plank2021opencarp] software focuses more on modeling the electrophysiology in the heart, while OpenCMISS[@bradley2011opencmiss] is more focused on cardiac mechanics.
+The OpenCarp [@plank2021opencarp] software focuses more on modeling the electrophysiology in the heart, while OpenCMISS [@bradley2011opencmiss] is more focused on cardiac mechanics.
 
-The availability of software for performing fully coupled cardiac electro-mechanics simulations is currently limited, and the ones that exist are not available as open source software and/or are typically written in a low-level language such as C[@arens2018gems] or C++[@Cooper2020]. The SIMula CARDiac Electro-Mechanics Solver, abbreviated `simcardems`, fuses the functionality from `pulse`[@finsberg2019pulse] and `cbcbeat`[@rognes2017cbcbeat] which are both based on the open source finite element framework FEniCS. The FEniCS library combines C++ and Python, providing an expressive language and highly efficient solvers that are made accessible to the broader scientific community through an intuitive interface in Python. FEniCS also offers a seamlessly parallel implementation with good scalability on a high performance computing cluster, enabling large scale realistic simulations.
[…]
