### 2023-01-09 #4753 @danielskatz
@finsberg - I've also created a PR (https://github.com/ComputationalPhysiology/simcardems/pull/132) with a few very minor changes - please merge this, or let me know what you disagree with
