### 2023-01-09 #4753 @danielskatz
@finsberg - I think this is ready to publish, once the [archive metadata issue](https://github.com/openjournals/joss-reviews/issues/4753#issuecomment-1375755825) is fixed. Please let me know when that is
