### 2023-01-09 #4753 @danielskatz
> @danielskatz I have now updated the title and the author list to match the paper (the original metadata was generated directly from a GitHub release).

@AoifeHughes - you might want add this in your instructions to future submitters.  See https://joss.readthedocs.io/en/latest/editing.html#message-to-authors-at-the-end-of-a-review for suggested language
